Move various AngelScript stuff out of the game module and into gameshared
