+++ source/game/ai/bot_movement.cpp
@@ -1059,11 +1059,11 @@ void BotMovementPredictionContext::BuildPlan() {
 		movementAction->BeforePlanning();
 
 	// Intercept these calls implicitly performed by PMove()
-	const auto general_PMoveTouchTriggers = module_PMoveTouchTriggers;
-	const auto general_PredictedEvent = module_PredictedEvent;
+	const auto general_PMoveTouchTriggers = gs.api.PMoveTouchTriggers;
+	const auto general_PredictedEvent = gs.api.PredictedEvent;
 
-	module_PMoveTouchTriggers = &Intercepted_PMoveTouchTriggers;
-	module_PredictedEvent = &Intercepted_PredictedEvent;
+	gs.api.PMoveTouchTriggers = &Intercepted_PMoveTouchTriggers;
+	gs.api.PredictedEvent = &Intercepted_PredictedEvent;
 
 	// The entity state might be modified by Intercepted_PMoveTouchTriggers(), so we have to save it
 	const Vec3 origin( self->s.origin );
@@ -1150,8 +1150,8 @@ void BotMovementPredictionContext::BuildPlan() {
 	Assert( VectorCompare( self->s.origin, self->ai->botRef->entityPhysicsState->Origin() ) );
 	Assert( VectorCompare( self->velocity, self->ai->botRef->entityPhysicsState->Velocity() ) );
 
-	module_PMoveTouchTriggers = general_PMoveTouchTriggers;
-	module_PredictedEvent = general_PredictedEvent;
+	gs.api.PMoveTouchTriggers = general_PMoveTouchTriggers;
+	gs.api.PredictedEvent = general_PredictedEvent;
 
 	for( auto *movementAction: self->ai->botRef->movementActions )
 		movementAction->AfterPlanning();
@@ -1210,30 +1210,30 @@ void BotMovementPredictionContext::NextMovementStep() {
 
 	// The naive solution of supplying a dummy trace function
 	// (that yields a zeroed output with fraction = 1) does not work.
-	// An actual logic tied to this flag has to be added in Pmove() for each module_Trace() call.
+	// An actual logic tied to this flag has to be added in Pmove() for each gs.api.Trace() call.
 	pm.skipCollision = EnvironmentTraceCache().CanSkipPMoveCollision( this );
 
 	// We currently test collisions only against a solid world on each movement step and the corresponding PMove() call.
 	// Touching trigger entities is handled by Intercepted_PMoveTouchTriggers(), also we use AAS sampling for it.
 	// Actions that involve touching trigger entities currently are never predicted ahead.
 	// If an action really needs to test against entities, a corresponding prediction step flag
-	// should be added and this interception of the module_Trace() should be skipped if the flag is set.
+	// should be added and this interception of the gs.api.Trace() should be skipped if the flag is set.
 
 	// Save the G_GS_Trace() pointer
-	auto oldModuleTrace = module_Trace;
-	module_Trace = Intercepted_Trace;
+	auto oldModuleTrace = gs.api.Trace;
+	gs.api.Trace = Intercepted_Trace;
 
 	// Do not test entities contents for same reasons
 	// Save the G_PointContents4D() pointer
-	auto oldModulePointContents = module_PointContents;
-	module_PointContents = Intercepted_PointContents;
+	auto oldModulePointContents = gs.api.PointContents;
+	gs.api.PointContents = Intercepted_PointContents;
 
 	Pmove( &pm );
 
 	// Restore the G_GS_Trace() pointer
-	module_Trace = oldModuleTrace;
+	gs.api.Trace = oldModuleTrace;
 	// Restore the G_PointContents4D() pointer
-	module_PointContents = oldModulePointContents;
+	gs.api.PointContents = oldModulePointContents;
 
 	// Update the entity physics state that is going to be used in the next prediction frame
 	entityPhysicsState->UpdateFromPMove( &pm );
