+++ source/game/g_as_local.h
@@ -20,11 +20,10 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "angelscript.h"
 #include "../gameshared/q_angeliface.h"
+#include "../gameshared/gs_ascript.h"
 
 #define G_AsMalloc                              G_LevelMalloc
 #define G_AsFree                                G_LevelFree
-
-extern angelwrap_api_t *angelExport;
 
 #define GAME_SCRIPTS_DIRECTORY              "progs"
 
@@ -39,82 +38,3 @@ extern angelwrap_api_t *angelExport;
 
 asIScriptModule *G_LoadGameScript( const char *moduleName, const char *dir, const char *filename, const char *ext );
 bool G_ExecutionErrorReport( int error );
-
-typedef struct asEnumVal_s {
-	const char * name;
-	int value;
-} asEnumVal_t;
-
-typedef struct asEnum_s {
-	const char * name;
-	const asEnumVal_t * values;
-} asEnum_t;
-
-typedef struct asFuncdef_s {
-	const char * declaration;
-} asFuncdef_t;
-
-typedef struct asBehavior_s {
-	asEBehaviours behavior;
-	const char * declaration;
-	asSFuncPtr funcPointer;
-	asECallConvTypes callConv;
-} asBehavior_t;
-
-typedef struct asMethod_s {
-	const char * declaration;
-	asSFuncPtr funcPointer;
-	asECallConvTypes callConv;
-} asMethod_t;
-
-typedef struct asProperty_s {
-	const char * declaration;
-	unsigned int offset;
-} asProperty_t;
-
-typedef struct asClassDescriptor_s {
-	const char * name;
-	asDWORD typeFlags;
-	size_t size;
-	const asFuncdef_t * funcdefs;
-	const asBehavior_t * objBehaviors;
-	const asMethod_t * objMethods;
-	const asProperty_t * objProperties;
-	const void * stringFactory;
-	const void * stringFactory_asGeneric;
-} asClassDescriptor_t;
-
-typedef struct asglobfuncs_s {
-	const char *declaration;
-	asSFuncPtr pointer;
-	void **asFuncPtr;
-} asglobfuncs_t;
-
-typedef struct asglobproperties_s {
-	const char *declaration;
-	void *pointer;
-} asglobproperties_t;
-
-static void asnullfunc( void ) {}
-
-#define ASLIB_LOCAL_CLASS_DESCR( x )
-
-#define ASLIB_FOFFSET( s,m )                      offsetof( s,m )
-
-#define ASLIB_ENUM_VAL( name )                    { #name,(int)name }
-#define ASLIB_ENUM_VAL_NULL                     { NULL, 0 }
-
-#define ASLIB_ENUM_NULL                         { NULL, NULL }
-
-#define ASLIB_FUNCTION_DECL( type,name,params )   (#type " " #name #params )
-
-#define ASLIB_PROPERTY_DECL( type,name )          #type " " #name
-
-#define ASLIB_FUNCTION_NULL                     NULL
-#define ASLIB_FUNCDEF_NULL                      { ASLIB_FUNCTION_NULL }
-#define ASLIB_BEHAVIOR_NULL                     { asBEHAVE_CONSTRUCT, ASLIB_FUNCTION_NULL, asFUNCTION( asnullfunc ), asCALL_CDECL }
-#define ASLIB_METHOD_NULL                       { ASLIB_FUNCTION_NULL, asFUNCTION( asnullfunc ), asCALL_CDECL }
-#define ASLIB_PROPERTY_NULL                     { NULL, 0 }
-
-#define ASLIB_Malloc( s )                         ( aslib_import.Mem_Alloc( s,__FILE__,__LINE__ ) )
-#define ASLIB_Free( x )                           ( aslib_import.Mem_Free( x,__FILE__,__LINE__ ) )
