+++ source/game/g_ascript.cpp
@@ -21,86 +21,9 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "g_local.h"
 #include "g_as_local.h"
 
-angelwrap_api_t *angelExport = NULL;
-
 //=======================================================================
 
-static const asEnumVal_t asConfigstringEnumVals[] =
-{
-	ASLIB_ENUM_VAL( CS_MODMANIFEST ),
-	ASLIB_ENUM_VAL( CS_MESSAGE ),
-	ASLIB_ENUM_VAL( CS_MAPNAME ),
-	ASLIB_ENUM_VAL( CS_AUDIOTRACK ),
-	ASLIB_ENUM_VAL( CS_HOSTNAME ),
-	ASLIB_ENUM_VAL( CS_TVSERVER ),
-	ASLIB_ENUM_VAL( CS_SKYBOX ),
-	ASLIB_ENUM_VAL( CS_STATNUMS ),
-	ASLIB_ENUM_VAL( CS_POWERUPEFFECTS ),
-	ASLIB_ENUM_VAL( CS_GAMETYPETITLE ),
-	ASLIB_ENUM_VAL( CS_GAMETYPENAME ),
-	ASLIB_ENUM_VAL( CS_GAMETYPEVERSION ),
-	ASLIB_ENUM_VAL( CS_GAMETYPEAUTHOR ),
-	ASLIB_ENUM_VAL( CS_AUTORECORDSTATE ),
-	ASLIB_ENUM_VAL( CS_SCB_PLAYERTAB_LAYOUT ),
-	ASLIB_ENUM_VAL( CS_SCB_PLAYERTAB_TITLES ),
-	ASLIB_ENUM_VAL( CS_TEAM_ALPHA_NAME ),
-	ASLIB_ENUM_VAL( CS_TEAM_BETA_NAME ),
-	ASLIB_ENUM_VAL( CS_MAXCLIENTS ),
-	ASLIB_ENUM_VAL( CS_MAPCHECKSUM ),
-	ASLIB_ENUM_VAL( CS_MATCHNAME ),
-	ASLIB_ENUM_VAL( CS_MATCHSCORE ),
-	ASLIB_ENUM_VAL( CS_ACTIVE_CALLVOTE ),
-
-	ASLIB_ENUM_VAL( CS_MODELS ),
-	ASLIB_ENUM_VAL( CS_SOUNDS ),
-	ASLIB_ENUM_VAL( CS_IMAGES ),
-	ASLIB_ENUM_VAL( CS_SKINFILES ),
-	ASLIB_ENUM_VAL( CS_LIGHTS ),
-	ASLIB_ENUM_VAL( CS_ITEMS ),
-	ASLIB_ENUM_VAL( CS_PLAYERINFOS ),
-	ASLIB_ENUM_VAL( CS_GAMECOMMANDS ),
-	ASLIB_ENUM_VAL( CS_LOCATIONS ),
-	ASLIB_ENUM_VAL( CS_GENERAL ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asEffectEnumVals[] =
-{
-	ASLIB_ENUM_VAL( EF_ROTATE_AND_BOB ),
-	ASLIB_ENUM_VAL( EF_SHELL ),
-	ASLIB_ENUM_VAL( EF_STRONG_WEAPON ),
-	ASLIB_ENUM_VAL( EF_QUAD ),
-	ASLIB_ENUM_VAL( EF_REGEN ),
-	ASLIB_ENUM_VAL( EF_CARRIER ),
-	ASLIB_ENUM_VAL( EF_BUSYICON ),
-	ASLIB_ENUM_VAL( EF_FLAG_TRAIL ),
-	ASLIB_ENUM_VAL( EF_TAKEDAMAGE ),
-	ASLIB_ENUM_VAL( EF_TEAMCOLOR_TRANSITION ),
-	ASLIB_ENUM_VAL( EF_EXPIRING_QUAD ),
-	ASLIB_ENUM_VAL( EF_EXPIRING_SHELL ),
-	ASLIB_ENUM_VAL( EF_EXPIRING_REGEN ),
-	ASLIB_ENUM_VAL( EF_GODMODE ),
-
-	ASLIB_ENUM_VAL( EF_PLAYER_STUNNED ),
-	ASLIB_ENUM_VAL( EF_PLAYER_HIDENAME ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asMatchStateEnumVals[] =
-{
-	//ASLIB_ENUM_VAL( MATCH_STATE_NONE ), // I see no point in adding it
-	ASLIB_ENUM_VAL( MATCH_STATE_WARMUP ),
-	ASLIB_ENUM_VAL( MATCH_STATE_COUNTDOWN ),
-	ASLIB_ENUM_VAL( MATCH_STATE_PLAYTIME ),
-	ASLIB_ENUM_VAL( MATCH_STATE_POSTMATCH ),
-	ASLIB_ENUM_VAL( MATCH_STATE_WAITEXIT ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asSpawnSystemEnumVals[] =
+static const gs_asEnumVal_t asSpawnSystemEnumVals[] =
 {
 	ASLIB_ENUM_VAL( SPAWNSYSTEM_INSTANT ),
 	ASLIB_ENUM_VAL( SPAWNSYSTEM_WAVES ),
@@ -109,85 +32,7 @@ static const asEnumVal_t asSpawnSystemEnumVals[] =
 	ASLIB_ENUM_VAL_NULL
 };
 
-static const asEnumVal_t asHUDStatEnumVals[] =
-{
-	ASLIB_ENUM_VAL( STAT_PROGRESS_SELF ),
-	ASLIB_ENUM_VAL( STAT_PROGRESS_OTHER ),
-	ASLIB_ENUM_VAL( STAT_PROGRESS_ALPHA ),
-	ASLIB_ENUM_VAL( STAT_PROGRESS_BETA ),
-	ASLIB_ENUM_VAL( STAT_IMAGE_SELF ),
-	ASLIB_ENUM_VAL( STAT_IMAGE_OTHER ),
-	ASLIB_ENUM_VAL( STAT_IMAGE_ALPHA ),
-	ASLIB_ENUM_VAL( STAT_IMAGE_BETA ),
-	ASLIB_ENUM_VAL( STAT_TIME_SELF ),
-	ASLIB_ENUM_VAL( STAT_TIME_BEST ),
-	ASLIB_ENUM_VAL( STAT_TIME_RECORD ),
-	ASLIB_ENUM_VAL( STAT_TIME_ALPHA ),
-	ASLIB_ENUM_VAL( STAT_TIME_BETA ),
-	ASLIB_ENUM_VAL( STAT_MESSAGE_SELF ),
-	ASLIB_ENUM_VAL( STAT_MESSAGE_OTHER ),
-	ASLIB_ENUM_VAL( STAT_MESSAGE_ALPHA ),
-	ASLIB_ENUM_VAL( STAT_MESSAGE_BETA ),
-	ASLIB_ENUM_VAL( STAT_IMAGE_CLASSACTION1 ),
-	ASLIB_ENUM_VAL( STAT_IMAGE_CLASSACTION2 ),
-	ASLIB_ENUM_VAL( STAT_IMAGE_DROP_ITEM ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asTeamEnumVals[] =
-{
-	ASLIB_ENUM_VAL( TEAM_SPECTATOR ),
-	ASLIB_ENUM_VAL( TEAM_PLAYERS ),
-	ASLIB_ENUM_VAL( TEAM_ALPHA ),
-	ASLIB_ENUM_VAL( TEAM_BETA ),
-	ASLIB_ENUM_VAL( GS_MAX_TEAMS ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asEntityTypeEnumVals[] =
-{
-	ASLIB_ENUM_VAL( ET_GENERIC ),
-	ASLIB_ENUM_VAL( ET_PLAYER ),
-	ASLIB_ENUM_VAL( ET_CORPSE ),
-	ASLIB_ENUM_VAL( ET_BEAM ),
-	ASLIB_ENUM_VAL( ET_PORTALSURFACE ),
-	ASLIB_ENUM_VAL( ET_PUSH_TRIGGER ),
-	ASLIB_ENUM_VAL( ET_GIB ),
-	ASLIB_ENUM_VAL( ET_BLASTER ),
-	ASLIB_ENUM_VAL( ET_ELECTRO_WEAK ),
-	ASLIB_ENUM_VAL( ET_ROCKET ),
-	ASLIB_ENUM_VAL( ET_GRENADE ),
-	ASLIB_ENUM_VAL( ET_PLASMA ),
-	ASLIB_ENUM_VAL( ET_SPRITE ),
-	ASLIB_ENUM_VAL( ET_ITEM ),
-	ASLIB_ENUM_VAL( ET_LASERBEAM ),
-	ASLIB_ENUM_VAL( ET_CURVELASERBEAM ),
-	ASLIB_ENUM_VAL( ET_FLAG_BASE ),
-	ASLIB_ENUM_VAL( ET_MINIMAP_ICON ),
-	ASLIB_ENUM_VAL( ET_DECAL ),
-	ASLIB_ENUM_VAL( ET_ITEM_TIMER ),
-	ASLIB_ENUM_VAL( ET_PARTICLES ),
-	ASLIB_ENUM_VAL( ET_SPAWN_INDICATOR ),
-	ASLIB_ENUM_VAL( ET_RADAR ),
-
-	ASLIB_ENUM_VAL( ET_EVENT ),
-	ASLIB_ENUM_VAL( ET_SOUNDEVENT ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asSolidEnumVals[] =
-{
-	ASLIB_ENUM_VAL( SOLID_NOT ),
-	ASLIB_ENUM_VAL( SOLID_TRIGGER ),
-	ASLIB_ENUM_VAL( SOLID_YES ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asMovetypeEnumVals[] =
+static const gs_asEnumVal_t asMovetypeEnumVals[] =
 {
 	ASLIB_ENUM_VAL( MOVETYPE_NONE ),
 	ASLIB_ENUM_VAL( MOVETYPE_PLAYER ),
@@ -204,286 +49,14 @@ static const asEnumVal_t asMovetypeEnumVals[] =
 	ASLIB_ENUM_VAL_NULL
 };
 
-static const asEnumVal_t asPMoveFeaturesVals[] =
-{
-	ASLIB_ENUM_VAL( PMFEAT_CROUCH ),
-	ASLIB_ENUM_VAL( PMFEAT_WALK ),
-	ASLIB_ENUM_VAL( PMFEAT_JUMP ),
-	ASLIB_ENUM_VAL( PMFEAT_DASH ),
-	ASLIB_ENUM_VAL( PMFEAT_WALLJUMP ),
-	ASLIB_ENUM_VAL( PMFEAT_FWDBUNNY ),
-	ASLIB_ENUM_VAL( PMFEAT_AIRCONTROL ),
-	ASLIB_ENUM_VAL( PMFEAT_ZOOM ),
-	ASLIB_ENUM_VAL( PMFEAT_GHOSTMOVE ),
-	ASLIB_ENUM_VAL( PMFEAT_CONTINOUSJUMP ),
-	ASLIB_ENUM_VAL( PMFEAT_ITEMPICK ),
-	ASLIB_ENUM_VAL( PMFEAT_GUNBLADEAUTOATTACK ),
-	ASLIB_ENUM_VAL( PMFEAT_WEAPONSWITCH ),
-	ASLIB_ENUM_VAL( PMFEAT_CORNERSKIMMING ),
-	ASLIB_ENUM_VAL( PMFEAT_CROUCHSLIDING ),
-	ASLIB_ENUM_VAL( PMFEAT_ALL ),
-	ASLIB_ENUM_VAL( PMFEAT_DEFAULT ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asItemTypeEnumVals[] =
-{
-	ASLIB_ENUM_VAL( IT_WEAPON ),
-	ASLIB_ENUM_VAL( IT_AMMO ),
-	ASLIB_ENUM_VAL( IT_ARMOR ),
-	ASLIB_ENUM_VAL( IT_POWERUP ),
-	ASLIB_ENUM_VAL( IT_HEALTH ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asInstagibNegItemMaskEnumVals[] =
+static const gs_asEnumVal_t asInstagibNegItemMaskEnumVals[] =
 {
 	ASLIB_ENUM_VAL( G_INSTAGIB_NEGATE_ITEMMASK ),
 
 	ASLIB_ENUM_VAL_NULL
 };
 
-static const asEnumVal_t asWeaponTagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( WEAP_NONE ),
-	ASLIB_ENUM_VAL( WEAP_GUNBLADE ),
-	ASLIB_ENUM_VAL( WEAP_MACHINEGUN ),
-	ASLIB_ENUM_VAL( WEAP_RIOTGUN ),
-	ASLIB_ENUM_VAL( WEAP_GRENADELAUNCHER ),
-	ASLIB_ENUM_VAL( WEAP_ROCKETLAUNCHER ),
-	ASLIB_ENUM_VAL( WEAP_PLASMAGUN ),
-	ASLIB_ENUM_VAL( WEAP_LASERGUN ),
-	ASLIB_ENUM_VAL( WEAP_ELECTROBOLT ),
-	ASLIB_ENUM_VAL( WEAP_INSTAGUN ),
-	ASLIB_ENUM_VAL( WEAP_TOTAL ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asAmmoTagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( AMMO_NONE ),
-	ASLIB_ENUM_VAL( AMMO_GUNBLADE ),
-	ASLIB_ENUM_VAL( AMMO_BULLETS ),
-	ASLIB_ENUM_VAL( AMMO_SHELLS ),
-	ASLIB_ENUM_VAL( AMMO_GRENADES ),
-	ASLIB_ENUM_VAL( AMMO_ROCKETS ),
-	ASLIB_ENUM_VAL( AMMO_PLASMA ),
-	ASLIB_ENUM_VAL( AMMO_LASERS ),
-	ASLIB_ENUM_VAL( AMMO_BOLTS ),
-	ASLIB_ENUM_VAL( AMMO_INSTAS ),
-
-	ASLIB_ENUM_VAL( AMMO_WEAK_GUNBLADE ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_BULLETS ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_SHELLS ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_GRENADES ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_ROCKETS ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_PLASMA ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_LASERS ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_BOLTS ),
-	ASLIB_ENUM_VAL( AMMO_WEAK_INSTAS ),
-
-	ASLIB_ENUM_VAL( AMMO_TOTAL ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asArmorTagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( ARMOR_NONE ),
-	ASLIB_ENUM_VAL( ARMOR_GA ),
-	ASLIB_ENUM_VAL( ARMOR_YA ),
-	ASLIB_ENUM_VAL( ARMOR_RA ),
-	ASLIB_ENUM_VAL( ARMOR_SHARD ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asHealthTagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( HEALTH_NONE ),
-	ASLIB_ENUM_VAL( HEALTH_SMALL ),
-	ASLIB_ENUM_VAL( HEALTH_MEDIUM ),
-	ASLIB_ENUM_VAL( HEALTH_LARGE ),
-	ASLIB_ENUM_VAL( HEALTH_MEGA ),
-	ASLIB_ENUM_VAL( HEALTH_ULTRA ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asPowerupTagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( POWERUP_NONE ),
-	ASLIB_ENUM_VAL( POWERUP_QUAD ),
-	ASLIB_ENUM_VAL( POWERUP_SHELL ),
-	ASLIB_ENUM_VAL( POWERUP_REGEN ),
-
-	ASLIB_ENUM_VAL( POWERUP_TOTAL ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asMiscItemTagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( AMMO_PACK_WEAK ),
-	ASLIB_ENUM_VAL( AMMO_PACK_STRONG ),
-	ASLIB_ENUM_VAL( AMMO_PACK ),
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asClientStateEnumVals[] =
-{
-	ASLIB_ENUM_VAL( CS_FREE ),
-	ASLIB_ENUM_VAL( CS_ZOMBIE ),
-	ASLIB_ENUM_VAL( CS_CONNECTING ),
-	ASLIB_ENUM_VAL( CS_CONNECTED ),
-	ASLIB_ENUM_VAL( CS_SPAWNED ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asSoundChannelEnumVals[] =
-{
-	ASLIB_ENUM_VAL( CHAN_AUTO ),
-	ASLIB_ENUM_VAL( CHAN_PAIN ),
-	ASLIB_ENUM_VAL( CHAN_VOICE ),
-	ASLIB_ENUM_VAL( CHAN_ITEM ),
-	ASLIB_ENUM_VAL( CHAN_BODY ),
-	ASLIB_ENUM_VAL( CHAN_MUZZLEFLASH ),
-	ASLIB_ENUM_VAL( CHAN_FIXED ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asContentsEnumVals[] =
-{
-	ASLIB_ENUM_VAL( CONTENTS_SOLID ),
-	ASLIB_ENUM_VAL( CONTENTS_LAVA ),
-	ASLIB_ENUM_VAL( CONTENTS_SLIME ),
-	ASLIB_ENUM_VAL( CONTENTS_WATER ),
-	ASLIB_ENUM_VAL( CONTENTS_FOG ),
-	ASLIB_ENUM_VAL( CONTENTS_AREAPORTAL ),
-	ASLIB_ENUM_VAL( CONTENTS_PLAYERCLIP ),
-	ASLIB_ENUM_VAL( CONTENTS_MONSTERCLIP ),
-	ASLIB_ENUM_VAL( CONTENTS_TELEPORTER ),
-	ASLIB_ENUM_VAL( CONTENTS_JUMPPAD ),
-	ASLIB_ENUM_VAL( CONTENTS_CLUSTERPORTAL ),
-	ASLIB_ENUM_VAL( CONTENTS_DONOTENTER ),
-	ASLIB_ENUM_VAL( CONTENTS_ORIGIN ),
-	ASLIB_ENUM_VAL( CONTENTS_BODY ),
-	ASLIB_ENUM_VAL( CONTENTS_CORPSE ),
-	ASLIB_ENUM_VAL( CONTENTS_DETAIL ),
-	ASLIB_ENUM_VAL( CONTENTS_STRUCTURAL ),
-	ASLIB_ENUM_VAL( CONTENTS_TRANSLUCENT ),
-	ASLIB_ENUM_VAL( CONTENTS_TRIGGER ),
-	ASLIB_ENUM_VAL( CONTENTS_NODROP ),
-	ASLIB_ENUM_VAL( MASK_ALL ),
-	ASLIB_ENUM_VAL( MASK_SOLID ),
-	ASLIB_ENUM_VAL( MASK_PLAYERSOLID ),
-	ASLIB_ENUM_VAL( MASK_DEADSOLID ),
-	ASLIB_ENUM_VAL( MASK_MONSTERSOLID ),
-	ASLIB_ENUM_VAL( MASK_WATER ),
-	ASLIB_ENUM_VAL( MASK_OPAQUE ),
-	ASLIB_ENUM_VAL( MASK_SHOT ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asSurfFlagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( SURF_NODAMAGE ),
-	ASLIB_ENUM_VAL( SURF_SLICK ),
-	ASLIB_ENUM_VAL( SURF_SKY ),
-	ASLIB_ENUM_VAL( SURF_LADDER ),
-	ASLIB_ENUM_VAL( SURF_NOIMPACT ),
-	ASLIB_ENUM_VAL( SURF_NOMARKS ),
-	ASLIB_ENUM_VAL( SURF_FLESH ),
-	ASLIB_ENUM_VAL( SURF_NODRAW ),
-	ASLIB_ENUM_VAL( SURF_HINT ),
-	ASLIB_ENUM_VAL( SURF_SKIP ),
-	ASLIB_ENUM_VAL( SURF_NOLIGHTMAP ),
-	ASLIB_ENUM_VAL( SURF_POINTLIGHT ),
-	ASLIB_ENUM_VAL( SURF_METALSTEPS ),
-	ASLIB_ENUM_VAL( SURF_NOSTEPS ),
-	ASLIB_ENUM_VAL( SURF_NONSOLID ),
-	ASLIB_ENUM_VAL( SURF_LIGHTFILTER ),
-	ASLIB_ENUM_VAL( SURF_ALPHASHADOW ),
-	ASLIB_ENUM_VAL( SURF_NODLIGHT ),
-	ASLIB_ENUM_VAL( SURF_DUST ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asSVFlagEnumVals[] =
-{
-	ASLIB_ENUM_VAL( SVF_NOCLIENT ),
-	ASLIB_ENUM_VAL( SVF_PORTAL ),
-	ASLIB_ENUM_VAL( SVF_TRANSMITORIGIN2 ),
-	ASLIB_ENUM_VAL( SVF_SOUNDCULL ),
-	ASLIB_ENUM_VAL( SVF_FAKECLIENT ),
-	ASLIB_ENUM_VAL( SVF_BROADCAST ),
-	ASLIB_ENUM_VAL( SVF_CORPSE ),
-	ASLIB_ENUM_VAL( SVF_PROJECTILE ),
-	ASLIB_ENUM_VAL( SVF_ONLYTEAM ),
-	ASLIB_ENUM_VAL( SVF_FORCEOWNER ),
-	ASLIB_ENUM_VAL( SVF_ONLYOWNER ),
-	ASLIB_ENUM_VAL( SVF_FORCETEAM ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asMeaningsOfDeathEnumVals[] =
-{
-	ASLIB_ENUM_VAL( MOD_GUNBLADE_W ),
-	ASLIB_ENUM_VAL( MOD_GUNBLADE_S ),
-	ASLIB_ENUM_VAL( MOD_MACHINEGUN_W ),
-	ASLIB_ENUM_VAL( MOD_MACHINEGUN_S ),
-	ASLIB_ENUM_VAL( MOD_RIOTGUN_W ),
-	ASLIB_ENUM_VAL( MOD_RIOTGUN_S ),
-	ASLIB_ENUM_VAL( MOD_GRENADE_W ),
-	ASLIB_ENUM_VAL( MOD_GRENADE_S ),
-	ASLIB_ENUM_VAL( MOD_ROCKET_W ),
-	ASLIB_ENUM_VAL( MOD_ROCKET_S ),
-	ASLIB_ENUM_VAL( MOD_PLASMA_W ),
-	ASLIB_ENUM_VAL( MOD_PLASMA_S ),
-	ASLIB_ENUM_VAL( MOD_ELECTROBOLT_W ),
-	ASLIB_ENUM_VAL( MOD_ELECTROBOLT_S ),
-	ASLIB_ENUM_VAL( MOD_INSTAGUN_W ),
-	ASLIB_ENUM_VAL( MOD_INSTAGUN_S ),
-	ASLIB_ENUM_VAL( MOD_LASERGUN_W ),
-	ASLIB_ENUM_VAL( MOD_LASERGUN_S ),
-	ASLIB_ENUM_VAL( MOD_GRENADE_SPLASH_W ),
-	ASLIB_ENUM_VAL( MOD_GRENADE_SPLASH_S ),
-	ASLIB_ENUM_VAL( MOD_ROCKET_SPLASH_W ),
-	ASLIB_ENUM_VAL( MOD_ROCKET_SPLASH_S ),
-	ASLIB_ENUM_VAL( MOD_PLASMA_SPLASH_W ),
-	ASLIB_ENUM_VAL( MOD_PLASMA_SPLASH_S ),
-
-	// World damage
-	ASLIB_ENUM_VAL( MOD_WATER ),
-	ASLIB_ENUM_VAL( MOD_SLIME ),
-	ASLIB_ENUM_VAL( MOD_LAVA ),
-	ASLIB_ENUM_VAL( MOD_CRUSH ),
-	ASLIB_ENUM_VAL( MOD_TELEFRAG ),
-	ASLIB_ENUM_VAL( MOD_FALLING ),
-	ASLIB_ENUM_VAL( MOD_SUICIDE ),
-	ASLIB_ENUM_VAL( MOD_EXPLOSIVE ),
-
-	// probably not used
-	ASLIB_ENUM_VAL( MOD_BARREL ),
-	ASLIB_ENUM_VAL( MOD_BOMB ),
-	ASLIB_ENUM_VAL( MOD_EXIT ),
-	ASLIB_ENUM_VAL( MOD_SPLASH ),
-	ASLIB_ENUM_VAL( MOD_TARGET_LASER ),
-	ASLIB_ENUM_VAL( MOD_TRIGGER_HURT ),
-	ASLIB_ENUM_VAL( MOD_HIT ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asDamageEnumVals[] =
+static const gs_asEnumVal_t asDamageEnumVals[] =
 {
 	ASLIB_ENUM_VAL( DAMAGE_NO ),
 	ASLIB_ENUM_VAL( DAMAGE_YES ),
@@ -492,295 +65,34 @@ static const asEnumVal_t asDamageEnumVals[] =
 	ASLIB_ENUM_VAL_NULL
 };
 
-static const asEnumVal_t asKeyiconEnumVals[] =
-{
-	ASLIB_ENUM_VAL( KEYICON_FORWARD ),
-	ASLIB_ENUM_VAL( KEYICON_BACKWARD ),
-	ASLIB_ENUM_VAL( KEYICON_LEFT ),
-	ASLIB_ENUM_VAL( KEYICON_RIGHT ),
-	ASLIB_ENUM_VAL( KEYICON_FIRE ),
-	ASLIB_ENUM_VAL( KEYICON_JUMP ),
-	ASLIB_ENUM_VAL( KEYICON_CROUCH ),
-	ASLIB_ENUM_VAL( KEYICON_SPECIAL ),
-	ASLIB_ENUM_VAL( KEYICON_TOTAL ),
-
-	ASLIB_ENUM_VAL_NULL
-};
-
-static const asEnumVal_t asMiscelaneaEnumVals[] =
+static const gs_asEnumVal_t asMiscelaneaEnumVals[] =
 {
 	ASLIB_ENUM_VAL_NULL
 };
 
 //=======================================================================
 
-static const asEnum_t asGameEnums[] =
+static const gs_asEnum_t asGameEnums[] =
 {
-	{ "configstrings_e", asConfigstringEnumVals },
-	{ "state_effects_e", asEffectEnumVals },
-	{ "matchstates_e", asMatchStateEnumVals },
 	{ "spawnsystem_e", asSpawnSystemEnumVals },
-	{ "hudstats_e", asHUDStatEnumVals },
-	{ "teams_e", asTeamEnumVals },
-	{ "entitytype_e", asEntityTypeEnumVals },
-	{ "solid_e", asSolidEnumVals },
 	{ "movetype_e", asMovetypeEnumVals },
-	{ "pmovefeats_e", asPMoveFeaturesVals },
-	{ "itemtype_e", asItemTypeEnumVals },
 
-	// we can't register defines, so we create a enum for each one of them :/
 	{ "G_INSTAGIB_NEGATE_ITEMMASK_e", asInstagibNegItemMaskEnumVals },
-	{ "weapon_tag_e", asWeaponTagEnumVals },
-	{ "ammo_tag_e", asAmmoTagEnumVals },
-	{ "armor_tag_e", asArmorTagEnumVals },
-	{ "health_tag_e", asHealthTagEnumVals },
-	{ "powerup_tag_e", asPowerupTagEnumVals },
-	{ "otheritems_tag_e", asMiscItemTagEnumVals },
 
-	{ "client_statest_e", asClientStateEnumVals },
-	{ "sound_channels_e", asSoundChannelEnumVals },
-	{ "contents_e", asContentsEnumVals },
-	{ "surfaceflags_e", asSurfFlagEnumVals },
-	{ "serverflags_e", asSVFlagEnumVals },
-	{ "meaningsofdeath_e", asMeaningsOfDeathEnumVals },
 	{ "takedamage_e", asDamageEnumVals },
-	{ "keyicon_e", asKeyiconEnumVals },
 	{ "miscelanea_e", asMiscelaneaEnumVals },
 
 	ASLIB_ENUM_VAL_NULL
 };
 
-/*
-* G_asRegisterEnums
-*/
-static void G_asRegisterEnums( asIScriptEngine *asEngine, const asEnum_t *asEnums ) {
-	int i, j;
-	const asEnum_t *asEnum;
-	const asEnumVal_t *asEnumVal;
-
-	for( i = 0, asEnum = asEnums; asEnum->name != NULL; i++, asEnum++ ) {
-		asEngine->RegisterEnum( asEnum->name );
-
-		for( j = 0, asEnumVal = asEnum->values; asEnumVal->name != NULL; j++, asEnumVal++ )
-			asEngine->RegisterEnumValue( asEnum->name, asEnumVal->name, asEnumVal->value );
-	}
-}
-
 //=======================================================================
 
 static asIObjectType *asEntityArrayType() {
-	asIScriptContext *ctx = angelExport->asGetActiveContext();
+	asIScriptContext *ctx = game.asExport->asGetActiveContext();
 	asIScriptEngine *engine = ctx->GetEngine();
 	asIObjectType *ot = engine->GetObjectTypeById( engine->GetTypeIdByDecl( "array<Entity @>" ) );
 	return ot;
 }
-
-//=======================================================================
-
-// CLASS: Trace
-typedef struct
-{
-	trace_t trace;
-} astrace_t;
-
-void objectTrace_DefaultConstructor( astrace_t *self ) {
-	memset( &self->trace, 0, sizeof( trace_t ) );
-}
-
-void objectTrace_CopyConstructor( astrace_t *other, astrace_t *self ) {
-	self->trace = other->trace;
-}
-
-static bool objectTrace_doTrace( asvec3_t *start, asvec3_t *mins, asvec3_t *maxs, asvec3_t *end, int ignore, int contentMask, astrace_t *self ) {
-	edict_t *passEnt = NULL;
-
-	if( ignore > 0 && ignore < game.maxentities ) {
-		passEnt = &game.edicts[ ignore ];
-	}
-
-	if( !start || !end ) { // should never happen unless the coder explicitly feeds null
-		G_Printf( "* WARNING: gametype plug-in script attempted to call method 'trace.doTrace' with a null vector pointer\n* Tracing skept" );
-		return false;
-	}
-
-	G_Trace( &self->trace, start->v, mins ? mins->v : vec3_origin, maxs ? maxs->v : vec3_origin, end->v, passEnt, contentMask );
-
-	if( self->trace.startsolid || self->trace.allsolid ) {
-		return true;
-	}
-
-	return ( self->trace.ent != -1 ) ? true : false;
-}
-
-static asvec3_t objectTrace_getEndPos( astrace_t *self ) {
-	asvec3_t asvec;
-
-	VectorCopy( self->trace.endpos, asvec.v );
-	return asvec;
-}
-
-static asvec3_t objectTrace_getPlaneNormal( astrace_t *self ) {
-	asvec3_t asvec;
-
-	VectorCopy( self->trace.plane.normal, asvec.v );
-	return asvec;
-}
-
-static const asFuncdef_t astrace_Funcdefs[] =
-{
-	ASLIB_FUNCDEF_NULL
-};
-
-static const asBehavior_t astrace_ObjectBehaviors[] =
-{
-	{ asBEHAVE_CONSTRUCT, ASLIB_FUNCTION_DECL( void, f, ( ) ), asFUNCTION( objectTrace_DefaultConstructor ), asCALL_CDECL_OBJLAST },
-	{ asBEHAVE_CONSTRUCT, ASLIB_FUNCTION_DECL( void, f, ( const Trace &in ) ), asFUNCTION( objectTrace_CopyConstructor ), asCALL_CDECL_OBJLAST },
-
-	ASLIB_BEHAVIOR_NULL
-};
-
-static const asMethod_t astrace_Methods[] =
-{
-	{ ASLIB_FUNCTION_DECL( bool, doTrace, ( const Vec3 &in, const Vec3 &in, const Vec3 &in, const Vec3 &in, int ignore, int contentMask ) const ), asFUNCTION( objectTrace_doTrace ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( Vec3, get_endPos, ( ) const ), asFUNCTION( objectTrace_getEndPos ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( Vec3, get_planeNormal, ( ) const ), asFUNCTION( objectTrace_getPlaneNormal ), asCALL_CDECL_OBJLAST },
-
-	ASLIB_METHOD_NULL
-};
-
-static const asProperty_t astrace_Properties[] =
-{
-	{ ASLIB_PROPERTY_DECL( const bool, allSolid ), ASLIB_FOFFSET( astrace_t, trace.allsolid ) },
-	{ ASLIB_PROPERTY_DECL( const bool, startSolid ), ASLIB_FOFFSET( astrace_t, trace.startsolid ) },
-	{ ASLIB_PROPERTY_DECL( const float, fraction ), ASLIB_FOFFSET( astrace_t, trace.fraction ) },
-	{ ASLIB_PROPERTY_DECL( const int, surfFlags ), ASLIB_FOFFSET( astrace_t, trace.surfFlags ) },
-	{ ASLIB_PROPERTY_DECL( const int, contents ), ASLIB_FOFFSET( astrace_t, trace.contents ) },
-	{ ASLIB_PROPERTY_DECL( const int, entNum ), ASLIB_FOFFSET( astrace_t, trace.ent ) },
-	{ ASLIB_PROPERTY_DECL( const float, planeDist ), ASLIB_FOFFSET( astrace_t, trace.plane.dist ) },
-	{ ASLIB_PROPERTY_DECL( const int16, planeType ), ASLIB_FOFFSET( astrace_t, trace.plane.type ) },
-	{ ASLIB_PROPERTY_DECL( const int16, planeSignBits ), ASLIB_FOFFSET( astrace_t, trace.plane.signbits ) },
-
-	ASLIB_PROPERTY_NULL
-};
-
-static const asClassDescriptor_t asTraceClassDescriptor =
-{
-	"Trace",                    /* name */
-	asOBJ_VALUE | asOBJ_POD | asOBJ_APP_CLASS_CK,   /* object type flags */
-	sizeof( astrace_t ),        /* size */
-	astrace_Funcdefs,           /* funcdefs */
-	astrace_ObjectBehaviors,    /* object behaviors */
-	astrace_Methods,            /* methods */
-	astrace_Properties,         /* properties */
-
-	NULL, NULL                  /* string factory hack */
-};
-
-//=======================================================================
-
-// CLASS: Item
-static asstring_t *objectGItem_getClassName( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->classname, self->classname ? strlen( self->classname ) : 0 );
-}
-
-static asstring_t *objectGItem_getName( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->name, self->name ? strlen( self->name ) : 0 );
-}
-
-static asstring_t *objectGItem_getShortName( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->shortname, self->shortname ? strlen( self->shortname ) : 0 );
-}
-
-static asstring_t *objectGItem_getModelName( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->world_model[0], self->world_model[0] ? strlen( self->world_model[0] ) : 0 );
-}
-
-static asstring_t *objectGItem_getModel2Name( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->world_model[1], self->world_model[1] ? strlen( self->world_model[1] ) : 0 );
-}
-
-static asstring_t *objectGItem_getIconName( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->icon, self->icon ? strlen( self->icon ) : 0 );
-}
-
-static asstring_t *objectGItem_getSimpleItemName( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->simpleitem, self->simpleitem ? strlen( self->simpleitem ) : 0 );
-}
-
-static asstring_t *objectGItem_getPickupSoundName( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->pickup_sound, self->pickup_sound ? strlen( self->pickup_sound ) : 0 );
-}
-
-static asstring_t *objectGItem_getColorToken( gsitem_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->color, self->color ? strlen( self->color ) : 0 );
-}
-
-static bool objectGItem_isPickable( gsitem_t *self ) {
-	return ( self && ( self->flags & ITFLAG_PICKABLE ) ) ? true : false;
-}
-
-static bool objectGItem_isUsable( gsitem_t *self ) {
-	return ( self && ( self->flags & ITFLAG_USABLE ) ) ? true : false;
-}
-
-static bool objectGItem_isDropable( gsitem_t *self ) {
-	return ( self && ( self->flags & ITFLAG_DROPABLE ) ) ? true : false;
-}
-
-static const asFuncdef_t asitem_Funcdefs[] =
-{
-	ASLIB_FUNCDEF_NULL
-};
-
-static const asBehavior_t asitem_ObjectBehaviors[] =
-{
-	ASLIB_BEHAVIOR_NULL
-};
-
-static const asMethod_t asitem_Methods[] =
-{
-	{ ASLIB_FUNCTION_DECL( const String @, get_classname, ( ) const ), asFUNCTION( objectGItem_getClassName ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_name, ( ) const ), asFUNCTION( objectGItem_getName ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_shortName, ( ) const ), asFUNCTION( objectGItem_getShortName ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_model, ( ) const ), asFUNCTION( objectGItem_getModelName ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_model2, ( ) const ), asFUNCTION( objectGItem_getModel2Name ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_icon, ( ) const ), asFUNCTION( objectGItem_getIconName ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_simpleIcon, ( ) const ), asFUNCTION( objectGItem_getSimpleItemName ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_pickupSound, ( ) const ), asFUNCTION( objectGItem_getPickupSoundName ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( const String @, get_colorToken, ( ) const ), asFUNCTION( objectGItem_getColorToken ), asCALL_CDECL_OBJLAST },
-
-	{ ASLIB_FUNCTION_DECL( bool, isPickable, ( ) const ), asFUNCTION( objectGItem_isPickable ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( bool, isUsable, ( ) const ), asFUNCTION( objectGItem_isUsable ), asCALL_CDECL_OBJLAST },
-	{ ASLIB_FUNCTION_DECL( bool, isDropable, ( ) const ), asFUNCTION( objectGItem_isDropable ), asCALL_CDECL_OBJLAST },
-
-	ASLIB_METHOD_NULL
-};
-
-static const asProperty_t asitem_Properties[] =
-{
-	{ ASLIB_PROPERTY_DECL( const int, tag ), ASLIB_FOFFSET( gsitem_t, tag ) },
-	{ ASLIB_PROPERTY_DECL( const uint, type ), ASLIB_FOFFSET( gsitem_t, type ) },
-	{ ASLIB_PROPERTY_DECL( const int, flags ), ASLIB_FOFFSET( gsitem_t, flags ) },
-	{ ASLIB_PROPERTY_DECL( const int, quantity ), ASLIB_FOFFSET( gsitem_t, quantity ) },
-	{ ASLIB_PROPERTY_DECL( const int, inventoryMax ), ASLIB_FOFFSET( gsitem_t, inventory_max ) },
-	{ ASLIB_PROPERTY_DECL( const int, ammoTag ), ASLIB_FOFFSET( gsitem_t, ammo_tag ) },
-	{ ASLIB_PROPERTY_DECL( const int, weakAmmoTag ), ASLIB_FOFFSET( gsitem_t, weakammo_tag ) },
-
-	ASLIB_PROPERTY_NULL
-};
-
-static const asClassDescriptor_t asItemClassDescriptor =
-{
-	"Item",                     /* name */
-	asOBJ_REF | asOBJ_NOCOUNT,    /* object type flags */
-	sizeof( gsitem_t ),         /* size */
-	asitem_Funcdefs,            /* funcdefs */
-	asitem_ObjectBehaviors,     /* object behaviors */
-	asitem_Methods,             /* methods */
-	asitem_Properties,          /* properties */
-
-	NULL, NULL                  /* string factory hack */
-};
 
 //=======================================================================
 
@@ -851,13 +163,13 @@ static int objectMatch_getState( match_t *self ) {
 static asstring_t *objectMatch_getName( match_t *self ) {
 	const char *s = trap_GetConfigString( CS_MATCHNAME );
 
-	return angelExport->asStringFactoryBuffer( s, strlen( s ) );
+	return game.asExport->asStringFactoryBuffer( s, strlen( s ) );
 }
 
 static asstring_t *objectMatch_getScore( match_t *self ) {
 	const char *s = trap_GetConfigString( CS_MATCHSCORE );
 
-	return angelExport->asStringFactoryBuffer( s, strlen( s ) );
+	return game.asExport->asStringFactoryBuffer( s, strlen( s ) );
 }
 
 static void objectMatch_setName( asstring_t *name, match_t *self ) {
@@ -880,17 +192,17 @@ static void objectMatch_setClockOverride( int64_t time, match_t *self ) {
 	gs.gameState.stats[GAMESTAT_CLOCKOVERRIDE] = time;
 }
 
-static const asFuncdef_t match_Funcdefs[] =
+static const gs_asFuncdef_t match_Funcdefs[] =
 {
 	ASLIB_FUNCDEF_NULL
 };
 
-static const asBehavior_t match_ObjectBehaviors[] =
+static const gs_asBehavior_t match_ObjectBehaviors[] =
 {
 	ASLIB_BEHAVIOR_NULL
 };
 
-static const asMethod_t match_Methods[] =
+static const gs_asMethod_t match_Methods[] =
 {
 	{ ASLIB_FUNCTION_DECL( void, launchState, (int state) const ), asFUNCTION( objectMatch_launchState ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( void, startAutorecord, ( ) const ), asFUNCTION( objectMatch_startAutorecord ), asCALL_CDECL_OBJLAST },
@@ -916,12 +228,12 @@ static const asMethod_t match_Methods[] =
 	ASLIB_METHOD_NULL
 };
 
-static const asProperty_t match_Properties[] =
+static const gs_asProperty_t match_Properties[] =
 {
 	ASLIB_PROPERTY_NULL
 };
 
-static const asClassDescriptor_t asMatchClassDescriptor =
+static const gs_asClassDescriptor_t asMatchClassDescriptor =
 {
 	"Match",                    /* name */
 	static_cast<asEObjTypeFlags>( asOBJ_REF | asOBJ_NOHANDLE ), /* object type flags */
@@ -941,7 +253,7 @@ static const asClassDescriptor_t asMatchClassDescriptor =
 static asstring_t *objectGametypeDescriptor_getTitle( gametype_descriptor_t *self ) {
 	const char *s = trap_GetConfigString( CS_GAMETYPETITLE );
 
-	return angelExport->asStringFactoryBuffer( s, strlen( s ) );
+	return game.asExport->asStringFactoryBuffer( s, strlen( s ) );
 }
 
 static void objectGametypeDescriptor_setTitle( asstring_t *other, gametype_descriptor_t *self ) {
@@ -953,13 +265,13 @@ static void objectGametypeDescriptor_setTitle( asstring_t *other, gametype_descr
 }
 
 static asstring_t *objectGametypeDescriptor_getName( gametype_descriptor_t *self ) {
-	return angelExport->asStringFactoryBuffer( gs.gametypeName, strlen( gs.gametypeName ) );
+	return game.asExport->asStringFactoryBuffer( gs.gametypeName, strlen( gs.gametypeName ) );
 }
 
 static asstring_t *objectGametypeDescriptor_getVersion( gametype_descriptor_t *self ) {
 	const char *s = trap_GetConfigString( CS_GAMETYPEVERSION );
 
-	return angelExport->asStringFactoryBuffer( s, strlen( s ) );
+	return game.asExport->asStringFactoryBuffer( s, strlen( s ) );
 }
 
 static void objectGametypeDescriptor_setVersion( asstring_t *other, gametype_descriptor_t *self ) {
@@ -973,7 +285,7 @@ static void objectGametypeDescriptor_setVersion( asstring_t *other, gametype_des
 static asstring_t *objectGametypeDescriptor_getAuthor( gametype_descriptor_t *self ) {
 	const char *s = trap_GetConfigString( CS_GAMETYPEAUTHOR );
 
-	return angelExport->asStringFactoryBuffer( s, strlen( s ) );
+	return game.asExport->asStringFactoryBuffer( s, strlen( s ) );
 }
 
 static void objectGametypeDescriptor_setAuthor( asstring_t *other, gametype_descriptor_t *self ) {
@@ -987,7 +299,7 @@ static void objectGametypeDescriptor_setAuthor( asstring_t *other, gametype_desc
 static asstring_t *objectGametypeDescriptor_getManifest( gametype_descriptor_t *self ) {
 	const char *s = trap_GetConfigString( CS_MODMANIFEST );
 
-	return angelExport->asStringFactoryBuffer( s, strlen( s ) );
+	return game.asExport->asStringFactoryBuffer( s, strlen( s ) );
 }
 
 static void objectGametypeDescriptor_SetTeamSpawnsystem( int team, int spawnsystem, int wave_time, int wave_maxcount, bool spectate_team, gametype_descriptor_t *self ) {
@@ -1010,17 +322,17 @@ static bool objectGametypeDescriptor_isInvidualGameType( gametype_descriptor_t *
 	return GS_InvidualGameType();
 }
 
-static const asFuncdef_t gametypedescr_Funcdefs[] =
+static const gs_asFuncdef_t gametypedescr_Funcdefs[] =
 {
 	ASLIB_FUNCDEF_NULL
 };
 
-static const asBehavior_t gametypedescr_ObjectBehaviors[] =
+static const gs_asBehavior_t gametypedescr_ObjectBehaviors[] =
 {
 	ASLIB_BEHAVIOR_NULL
 };
 
-static const asMethod_t gametypedescr_Methods[] =
+static const gs_asMethod_t gametypedescr_Methods[] =
 {
 	{ ASLIB_FUNCTION_DECL( const String @, get_name, ( ) const ), asFUNCTION( objectGametypeDescriptor_getName ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( const String @, get_title, ( ) const ), asFUNCTION( objectGametypeDescriptor_getTitle ), asCALL_CDECL_OBJLAST },
@@ -1039,7 +351,7 @@ static const asMethod_t gametypedescr_Methods[] =
 	ASLIB_METHOD_NULL
 };
 
-static const asProperty_t gametypedescr_Properties[] =
+static const gs_asProperty_t gametypedescr_Properties[] =
 {
 	{ ASLIB_PROPERTY_DECL( uint, spawnableItemsMask ), ASLIB_FOFFSET( gametype_descriptor_t, spawnableItemsMask ) },
 	{ ASLIB_PROPERTY_DECL( uint, respawnableItemsMask ), ASLIB_FOFFSET( gametype_descriptor_t, respawnableItemsMask ) },
@@ -1082,7 +394,7 @@ static const asProperty_t gametypedescr_Properties[] =
 	ASLIB_PROPERTY_NULL
 };
 
-static const asClassDescriptor_t asGametypeClassDescriptor =
+static const gs_asClassDescriptor_t asGametypeClassDescriptor =
 {
 	"GametypeDesc",                 /* name */
 	asOBJ_REF | asOBJ_NOHANDLE,       /* object type flags */
@@ -1113,13 +425,13 @@ static edict_t *objectTeamlist_GetPlayerEntity( int index, g_teamlist_t *obj ) {
 static asstring_t *objectTeamlist_getName( g_teamlist_t *obj ) {
 	const char *name = GS_TeamName( obj - teamlist );
 
-	return angelExport->asStringFactoryBuffer( name, name ? strlen( name ) : 0 );
+	return game.asExport->asStringFactoryBuffer( name, name ? strlen( name ) : 0 );
 }
 
 static asstring_t *objectTeamlist_getDefaultName( g_teamlist_t *obj ) {
 	const char *name = GS_DefaultTeamName( obj - teamlist );
 
-	return angelExport->asStringFactoryBuffer( name, name ? strlen( name ) : 0 );
+	return game.asExport->asStringFactoryBuffer( name, name ? strlen( name ) : 0 );
 }
 
 static void objectTeamlist_setName( asstring_t *str, g_teamlist_t *obj ) {
@@ -1162,17 +474,17 @@ static int objectTeamlist_getTeamIndex( g_teamlist_t *obj ) {
 	return index;
 }
 
-static const asFuncdef_t teamlist_Funcdefs[] =
+static const gs_asFuncdef_t teamlist_Funcdefs[] =
 {
 	ASLIB_FUNCDEF_NULL
 };
 
-static const asBehavior_t teamlist_ObjectBehaviors[] =
+static const gs_asBehavior_t teamlist_ObjectBehaviors[] =
 {
 	ASLIB_BEHAVIOR_NULL
 };
 
-static const asMethod_t teamlist_Methods[] =
+static const gs_asMethod_t teamlist_Methods[] =
 {
 	{ ASLIB_FUNCTION_DECL( Entity @, ent, ( int index ) ), asFUNCTION( objectTeamlist_GetPlayerEntity ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( const String @, get_name, ( ) const ), asFUNCTION( objectTeamlist_getName ), asCALL_CDECL_OBJLAST },
@@ -1187,7 +499,7 @@ static const asMethod_t teamlist_Methods[] =
 	ASLIB_METHOD_NULL
 };
 
-static const asProperty_t teamlist_Properties[] =
+static const gs_asProperty_t teamlist_Properties[] =
 {
 	{ ASLIB_PROPERTY_DECL( Stats, stats ), ASLIB_FOFFSET( g_teamlist_t, stats ) },
 	{ ASLIB_PROPERTY_DECL( const int, numPlayers ), ASLIB_FOFFSET( g_teamlist_t, numplayers ) },
@@ -1197,7 +509,7 @@ static const asProperty_t teamlist_Properties[] =
 	ASLIB_PROPERTY_NULL
 };
 
-static const asClassDescriptor_t asTeamListClassDescriptor =
+static const gs_asClassDescriptor_t asTeamListClassDescriptor =
 {
 	"Team",                     /* name */
 	asOBJ_REF | asOBJ_NOCOUNT,    /* object type flags */
@@ -1269,17 +581,17 @@ static void objectScoreStats_RoundAdd( score_stats_t *obj ) {
 	obj->numrounds++;
 }
 
-static const asFuncdef_t scorestats_Funcdefs[] =
+static const gs_asFuncdef_t scorestats_Funcdefs[] =
 {
 	ASLIB_FUNCDEF_NULL
 };
 
-static const asBehavior_t scorestats_ObjectBehaviors[] =
+static const gs_asBehavior_t scorestats_ObjectBehaviors[] =
 {
 	ASLIB_BEHAVIOR_NULL
 };
 
-static const asMethod_t scorestats_Methods[] =
+static const gs_asMethod_t scorestats_Methods[] =
 {
 	{ ASLIB_FUNCTION_DECL( void, setScore, ( int i ) ), asFUNCTION( objectScoreStats_ScoreSet ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( void, addScore, ( int i ) ), asFUNCTION( objectScoreStats_ScoreAdd ), asCALL_CDECL_OBJLAST },
@@ -1294,7 +606,7 @@ static const asMethod_t scorestats_Methods[] =
 	ASLIB_METHOD_NULL
 };
 
-static const asProperty_t scorestats_Properties[] =
+static const gs_asProperty_t scorestats_Properties[] =
 {
 	{ ASLIB_PROPERTY_DECL( const int, score ), ASLIB_FOFFSET( score_stats_t, score ) },
 	{ ASLIB_PROPERTY_DECL( const int, deaths ), ASLIB_FOFFSET( score_stats_t, deaths ) },
@@ -1312,7 +624,7 @@ static const asProperty_t scorestats_Properties[] =
 	ASLIB_PROPERTY_NULL
 };
 
-static const asClassDescriptor_t asScoreStatsClassDescriptor =
+static const gs_asClassDescriptor_t asScoreStatsClassDescriptor =
 {
 	"Stats",                    /* name */
 	asOBJ_REF | asOBJ_NOCOUNT,    /* object type flags */
@@ -1393,7 +705,7 @@ static asstring_t *objectGameClient_getName( gclient_t *self ) {
 	Q_strncpyz( temp, self->netname, sizeof( temp ) );
 	Q_strncatz( temp, S_COLOR_WHITE, sizeof( temp ) );
 
-	return angelExport->asStringFactoryBuffer( temp, strlen( temp ) );
+	return game.asExport->asStringFactoryBuffer( temp, strlen( temp ) );
 }
 
 static asstring_t *objectGameClient_getClanName( gclient_t *self ) {
@@ -1402,7 +714,7 @@ static asstring_t *objectGameClient_getClanName( gclient_t *self ) {
 	Q_strncpyz( temp, self->clanname, sizeof( temp ) );
 	Q_strncatz( temp, S_COLOR_WHITE, sizeof( temp ) );
 
-	return angelExport->asStringFactoryBuffer( temp, strlen( temp ) );
+	return game.asExport->asStringFactoryBuffer( temp, strlen( temp ) );
 }
 
 static asstring_t *objectGameClient_getMMLogin( gclient_t *self ) {
@@ -1415,7 +727,7 @@ static asstring_t *objectGameClient_getMMLogin( gclient_t *self ) {
 		login = "";
 	}
 
-	return angelExport->asStringFactoryBuffer( login, strlen( login ) );
+	return game.asExport->asStringFactoryBuffer( login, strlen( login ) );
 }
 
 static void objectGameClient_Respawn( bool ghost, gclient_t *self ) {
@@ -1649,15 +961,15 @@ static asstring_t *objectGameClient_getUserInfoKey( asstring_t *key, gclient_t *
 	char *s;
 
 	if( !key || !key->buffer || !key->buffer[0] ) {
-		return angelExport->asStringFactoryBuffer( NULL, 0 );
+		return game.asExport->asStringFactoryBuffer( NULL, 0 );
 	}
 
 	s = Info_ValueForKey( self->userinfo, key->buffer );
 	if( !s || !*s ) {
-		return angelExport->asStringFactoryBuffer( NULL, 0 );
+		return game.asExport->asStringFactoryBuffer( NULL, 0 );
 	}
 
-	return angelExport->asStringFactoryBuffer( s, strlen( s ) );
+	return game.asExport->asStringFactoryBuffer( s, strlen( s ) );
 }
 
 static void objectGameClient_printMessage( asstring_t *str, gclient_t *self ) {
@@ -1755,17 +1067,17 @@ static void objectGameClient_SetQuickMenuItems( asstring_t *str, gclient_t *self
 	trap_GameCmd( PLAYERENT( playerNum ), va( "qm %s", str->buffer ) );
 }
 
-static const asFuncdef_t gameclient_Funcdefs[] =
+static const gs_asFuncdef_t gameclient_Funcdefs[] =
 {
 	ASLIB_FUNCDEF_NULL
 };
 
-static const asBehavior_t gameclient_ObjectBehaviors[] =
+static const gs_asBehavior_t gameclient_ObjectBehaviors[] =
 {
 	ASLIB_BEHAVIOR_NULL
 };
 
-static const asMethod_t gameclient_Methods[] =
+static const gs_asMethod_t gameclient_Methods[] =
 {
 	{ ASLIB_FUNCTION_DECL( int, get_playerNum, ( ) const ), asFUNCTION( objectGameClient_PlayerNum ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( bool, isReady, ( ) const ), asFUNCTION( objectGameClient_isReady ), asCALL_CDECL_OBJLAST },
@@ -1812,7 +1124,7 @@ static const asMethod_t gameclient_Methods[] =
 	ASLIB_METHOD_NULL
 };
 
-static const asProperty_t gameclient_Properties[] =
+static const gs_asProperty_t gameclient_Properties[] =
 {
 	{ ASLIB_PROPERTY_DECL( Stats, stats ), ASLIB_FOFFSET( gclient_t, level.stats ) },
 	{ ASLIB_PROPERTY_DECL( const bool, connecting ), ASLIB_FOFFSET( gclient_t, connecting ) },
@@ -1839,7 +1151,7 @@ static const asProperty_t gameclient_Properties[] =
 	ASLIB_PROPERTY_NULL
 };
 
-static const asClassDescriptor_t asGameClientDescriptor =
+static const gs_asClassDescriptor_t asGameClientDescriptor =
 {
 	"Client",                   /* name */
 	asOBJ_REF | asOBJ_NOCOUNT,    /* object type flags */
@@ -1978,15 +1290,15 @@ static int objectGameEntity_PlayerNum( edict_t *self ) {
 }
 
 static asstring_t *objectGameEntity_getModelName( edict_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->model, self->model ? strlen( self->model ) : 0 );
+	return game.asExport->asStringFactoryBuffer( self->model, self->model ? strlen( self->model ) : 0 );
 }
 
 static asstring_t *objectGameEntity_getModel2Name( edict_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->model2, self->model2 ? strlen( self->model2 ) : 0 );
+	return game.asExport->asStringFactoryBuffer( self->model2, self->model2 ? strlen( self->model2 ) : 0 );
 }
 
 static asstring_t *objectGameEntity_getClassname( edict_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->classname, self->classname ? strlen( self->classname ) : 0 );
+	return game.asExport->asStringFactoryBuffer( self->classname, self->classname ? strlen( self->classname ) : 0 );
 }
 
 /*
@@ -1995,16 +1307,16 @@ static asstring_t *objectGameEntity_getSpawnKey( asstring_t *key, edict_t *self 
 const char *val;
 
 if( !key )
-return angelExport->asStringFactoryBuffer( NULL, 0 );
+return game.asExport->asStringFactoryBuffer( NULL, 0 );
 
 val = G_GetEntitySpawnKey( key->buffer, self );
 
-return angelExport->asStringFactoryBuffer( val, strlen( val ) );
+return game.asExport->asStringFactoryBuffer( val, strlen( val ) );
 }
 */
 
 static asstring_t *objectGameEntity_getTargetname( edict_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->targetname, self->targetname ? strlen( self->targetname ) : 0 );
+	return game.asExport->asStringFactoryBuffer( self->targetname, self->targetname ? strlen( self->targetname ) : 0 );
 }
 
 static void objectGameEntity_setTargetname( asstring_t *targetname, edict_t *self ) {
@@ -2012,7 +1324,7 @@ static void objectGameEntity_setTargetname( asstring_t *targetname, edict_t *sel
 }
 
 static asstring_t *objectGameEntity_getTarget( edict_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->target, self->target ? strlen( self->target ) : 0 );
+	return game.asExport->asStringFactoryBuffer( self->target, self->target ? strlen( self->target ) : 0 );
 }
 
 static void objectGameEntity_setTarget( asstring_t *target, edict_t *self ) {
@@ -2020,11 +1332,11 @@ static void objectGameEntity_setTarget( asstring_t *target, edict_t *self ) {
 }
 
 static asstring_t *objectGameEntity_getMap( edict_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->map, self->map ? strlen( self->map ) : 0 );
+	return game.asExport->asStringFactoryBuffer( self->map, self->map ? strlen( self->map ) : 0 );
 }
 
 static asstring_t *objectGameEntity_getSoundName( edict_t *self ) {
-	return angelExport->asStringFactoryBuffer( self->sounds, self->sounds ? strlen( self->sounds ) : 0 );
+	return game.asExport->asStringFactoryBuffer( self->sounds, self->sounds ? strlen( self->sounds ) : 0 );
 }
 
 static void objectGameEntity_setClassname( asstring_t *classname, edict_t *self ) {
@@ -2099,7 +1411,7 @@ static edict_t *objectGameEntity_DropItem( gsitem_t *item, edict_t *self ) {
 
 static CScriptArrayInterface *objectGameEntity_findTargets( edict_t *self ) {
 	asIObjectType *ot = asEntityArrayType();
-	CScriptArrayInterface *arr = angelExport->asCreateArrayCpp( 0, ot );
+	CScriptArrayInterface *arr = game.asExport->asCreateArrayCpp( 0, ot );
 
 	if( self->target && self->target[0] != '\0' ) {
 		int count = 0;
@@ -2116,7 +1428,7 @@ static CScriptArrayInterface *objectGameEntity_findTargets( edict_t *self ) {
 
 static CScriptArrayInterface *objectGameEntity_findTargeting( edict_t *self ) {
 	asIObjectType *ot = asEntityArrayType();
-	CScriptArrayInterface *arr = angelExport->asCreateArrayCpp( 0, ot );
+	CScriptArrayInterface *arr = game.asExport->asCreateArrayCpp( 0, ot );
 
 	if( self->targetname && self->targetname[0] != '\0' ) {
 		int count = 0;
@@ -2184,7 +1496,7 @@ static void objectGameEntity_explosionEffect( int radius, edict_t *self ) {
 	G_SpawnEvent( eventType, eventRadius, center );
 }
 
-static const asFuncdef_t gedict_Funcdefs[] =
+static const gs_asFuncdef_t gedict_Funcdefs[] =
 {
 	{ ASLIB_FUNCTION_DECL( void, entThink, ( Entity @ent ) ) },
 	{ ASLIB_FUNCTION_DECL( void, entTouch, ( Entity @ent, Entity @other, const Vec3 planeNormal, int surfFlags ) ) },
@@ -2196,12 +1508,12 @@ static const asFuncdef_t gedict_Funcdefs[] =
 	ASLIB_FUNCDEF_NULL
 };
 
-static const asBehavior_t gedict_ObjectBehaviors[] =
+static const gs_asBehavior_t gedict_ObjectBehaviors[] =
 {
 	ASLIB_BEHAVIOR_NULL
 };
 
-static const asMethod_t gedict_Methods[] =
+static const gs_asMethod_t gedict_Methods[] =
 {
 	{ ASLIB_FUNCTION_DECL( Vec3, get_velocity, ( ) const ), asFUNCTION( objectGameEntity_GetVelocity ), asCALL_CDECL_OBJLAST },
 	{ ASLIB_FUNCTION_DECL( void, set_velocity, ( const Vec3 &in ) ), asFUNCTION( objectGameEntity_SetVelocity ), asCALL_CDECL_OBJLAST },
@@ -2254,7 +1566,7 @@ static const asMethod_t gedict_Methods[] =
 	ASLIB_METHOD_NULL
 };
 
-static const asProperty_t gedict_Properties[] =
+static const gs_asProperty_t gedict_Properties[] =
 {
 	{ ASLIB_PROPERTY_DECL( Client @, client ), ASLIB_FOFFSET( edict_t, r.client ) },
 	{ ASLIB_PROPERTY_DECL( Item @, item ), ASLIB_FOFFSET( edict_t, item ) },
@@ -2329,7 +1641,7 @@ static const asProperty_t gedict_Properties[] =
 	ASLIB_PROPERTY_NULL
 };
 
-static const asClassDescriptor_t asGameEntityClassDescriptor =
+static const gs_asClassDescriptor_t asGameEntityClassDescriptor =
 {
 	"Entity",                   /* name */
 	asOBJ_REF | asOBJ_NOCOUNT,    /* object type flags */
@@ -2344,10 +1656,8 @@ static const asClassDescriptor_t asGameEntityClassDescriptor =
 
 //=======================================================================
 
-static const asClassDescriptor_t * const asGameClassesDescriptors[] =
+static const gs_asClassDescriptor_t * const asGameClassesDescriptors[] =
 {
-	&asTraceClassDescriptor,
-	&asItemClassDescriptor,
 	&asMatchClassDescriptor,
 	&asGametypeClassDescriptor,
 	&asTeamListClassDescriptor,
@@ -2357,84 +1667,6 @@ static const asClassDescriptor_t * const asGameClassesDescriptors[] =
 
 	NULL
 };
-
-static void G_asRegisterObjectClassNames( asIScriptEngine *asEngine, const asClassDescriptor_t *const *asClassesDescriptors ) {
-	int i;
-	const asClassDescriptor_t *cDescr;
-
-	for( i = 0; ; i++ ) {
-		if( !( cDescr = asClassesDescriptors[i] ) ) {
-			break;
-		}
-		asEngine->RegisterObjectType( cDescr->name, cDescr->size, cDescr->typeFlags );
-	}
-}
-
-/*
-* G_asRegisterObjectClasses
-*/
-static void G_asRegisterObjectClasses( asIScriptEngine *asEngine, const asClassDescriptor_t *const *asClassesDescriptors ) {
-	int i, j;
-	const asClassDescriptor_t *cDescr;
-
-	// now register object and global behaviors, then methods and properties
-	for( i = 0; ; i++ ) {
-		if( !( cDescr = asClassesDescriptors[i] ) ) {
-			break;
-		}
-
-		// funcdefs
-		if( cDescr->funcdefs ) {
-			for( j = 0; ; j++ ) {
-				const asFuncdef_t *funcdef = &cDescr->funcdefs[j];
-				if( !funcdef->declaration ) {
-					break;
-				}
-				asEngine->RegisterFuncdef( funcdef->declaration );
-			}
-		}
-
-		// object behaviors
-		if( cDescr->objBehaviors ) {
-			for( j = 0; ; j++ ) {
-				const asBehavior_t *objBehavior = &cDescr->objBehaviors[j];
-				if( !objBehavior->declaration ) {
-					break;
-				}
-				asEngine->RegisterObjectBehaviour(
-					cDescr->name, objBehavior->behavior, objBehavior->declaration,
-					objBehavior->funcPointer, objBehavior->callConv );
-			}
-		}
-
-		// object methods
-		if( cDescr->objMethods ) {
-			for( j = 0; ; j++ ) {
-				const asMethod_t *objMethod = &cDescr->objMethods[j];
-				if( !objMethod->declaration ) {
-					break;
-				}
-
-				asEngine->RegisterObjectMethod( cDescr->name,
-												objMethod->declaration, objMethod->funcPointer,
-												objMethod->callConv );
-			}
-		}
-
-		// object properties
-		if( cDescr->objProperties ) {
-			for( j = 0; ; j++ ) {
-				const asProperty_t *objProperty = &cDescr->objProperties[j];
-				if( !objProperty->declaration ) {
-					break;
-				}
-
-				asEngine->RegisterObjectProperty( cDescr->name,
-												  objProperty->declaration, objProperty->offset );
-			}
-		}
-	}
-}
 
 //=======================================================================
 
@@ -2454,7 +1686,7 @@ static edict_t *asFunc_G_Spawn( asstring_t *classname ) {
 
 	ent->scriptSpawned = true;
 	ent->asScriptModule = static_cast<void *>(
-		GAME_AS_ENGINE()->GetModule( angelExport->asGetActiveContext()->GetFunction( 0 )->GetModuleName() )
+		GAME_AS_ENGINE()->GetModule( game.asExport->asGetActiveContext()->GetFunction( 0 )->GetModuleName() )
 		);
 
 	G_asClearEntityBehaviors( ent );
@@ -2648,7 +1880,7 @@ static asstring_t *asFunc_LoadFile( asstring_t *path ) {
 	asstring_t *data;
 
 	if( !path || !path->len ) {
-		return angelExport->asStringFactoryBuffer( NULL, 0 );
+		return game.asExport->asStringFactoryBuffer( NULL, 0 );
 	}
 
 	filelen = trap_FS_FOpenFile( path->buffer, &filehandle, FS_READ );
@@ -2660,10 +1892,10 @@ static asstring_t *asFunc_LoadFile( asstring_t *path ) {
 	trap_FS_FCloseFile( filehandle );
 
 	if( !buf ) {
-		return angelExport->asStringFactoryBuffer( NULL, 0 );
+		return game.asExport->asStringFactoryBuffer( NULL, 0 );
 	}
 
-	data = angelExport->asStringFactoryBuffer( (char *)buf, filelen );
+	data = game.asExport->asStringFactoryBuffer( (char *)buf, filelen );
 	G_Free( buf );
 
 	return data;
@@ -2697,7 +1929,7 @@ static asstring_t *asFunc_ML_GetMapByNum( int num ) {
 		return NULL;
 	}
 
-	data = angelExport->asStringFactoryBuffer( (char *)mapname, strlen( mapname ) );
+	data = game.asExport->asStringFactoryBuffer( (char *)mapname, strlen( mapname ) );
 	return data;
 }
 
@@ -2707,7 +1939,7 @@ static asstring_t *asFunc_LocationName( asvec3_t *origin ) {
 
 	G_MapLocationNameForTAG( G_MapLocationTAGForOrigin( origin->v ), buf, sizeof( buf ) );
 
-	data = angelExport->asStringFactoryBuffer( (char *)buf, strlen( buf ) );
+	data = game.asExport->asStringFactoryBuffer( (char *)buf, strlen( buf ) );
 	return data;
 }
 
@@ -2721,7 +1953,7 @@ static asstring_t *asFunc_LocationForTag( int tag ) {
 
 	G_MapLocationNameForTAG( tag, buf, sizeof( buf ) );
 
-	data = angelExport->asStringFactoryBuffer( (char *)buf, strlen( buf ) );
+	data = game.asExport->asStringFactoryBuffer( (char *)buf, strlen( buf ) );
 
 	return data;
 }
@@ -2801,7 +2033,7 @@ static void asFunc_RegisterCallvote( asstring_t *asname, asstring_t *asusage, as
 
 static asstring_t *asFunc_GetConfigString( int index ) {
 	const char *cs = trap_GetConfigString( index );
-	return angelExport->asStringFactoryBuffer( (char *)cs, cs ? strlen( cs ) : 0 );
+	return game.asExport->asStringFactoryBuffer( (char *)cs, cs ? strlen( cs ) : 0 );
 }
 
 static void asFunc_SetConfigString( int index, asstring_t *str ) {
@@ -2883,7 +2115,7 @@ static CScriptArrayInterface *asFunc_G_FindInRadius( asvec3_t *org, float radius
 
 	int touch[MAX_EDICTS];
 	int numtouch = GClip_FindInRadius( org->v, radius, touch, MAX_EDICTS );
-	CScriptArrayInterface *arr = angelExport->asCreateArrayCpp( numtouch, ot );
+	CScriptArrayInterface *arr = game.asExport->asCreateArrayCpp( numtouch, ot );
 	for( int i = 0; i < numtouch; i++ ) {
 		*( (edict_t **)arr->At( i ) ) = game.edicts + touch[i];
 	}
@@ -2895,7 +2127,7 @@ static CScriptArrayInterface *asFunc_G_FindByClassname( asstring_t *str ) {
 	const char *classname = str->buffer;
 
 	asIObjectType *ot = asEntityArrayType();
-	CScriptArrayInterface *arr = angelExport->asCreateArrayCpp( 0, ot );
+	CScriptArrayInterface *arr = game.asExport->asCreateArrayCpp( 0, ot );
 
 	int count = 0;
 	edict_t *ent = NULL;
@@ -2972,7 +2204,7 @@ static asstring_t *asFunc_G_SpawnTempValue( asstring_t *key ) {
 	const char *val;
 
 	if( !key ) {
-		return angelExport->asStringFactoryBuffer( NULL, 0 );
+		return game.asExport->asStringFactoryBuffer( NULL, 0 );
 	}
 
 	if( level.spawning_entity == NULL ) {
@@ -2981,7 +2213,7 @@ static asstring_t *asFunc_G_SpawnTempValue( asstring_t *key ) {
 
 	val = G_GetEntitySpawnKey( key->buffer, level.spawning_entity );
 
-	return angelExport->asStringFactoryBuffer( val, strlen( val ) );
+	return game.asExport->asStringFactoryBuffer( val, strlen( val ) );
 }
 
 static void asFunc_FireInstaShot( asvec3_t *origin, asvec3_t *angles, int range, int damage, int knockback, int stun, edict_t *owner ) {
@@ -3032,7 +2264,7 @@ static int asFunc_G_GetDefaultColorCorrection( void ) {
 	return level.colorCorrection;
 }
 
-static const asglobfuncs_t asGameGlobFuncs[] =
+static const gs_asglobfuncs_t asGameGlobFuncs[] =
 {
 	{ "Entity @G_SpawnEntity( const String &in )", asFUNCTION( asFunc_G_Spawn ), NULL },
 	{ "const String @G_SpawnTempValue( const String &in )", asFUNCTION( asFunc_G_SpawnTempValue ), NULL },
@@ -3129,7 +2361,7 @@ static const asglobfuncs_t asGameGlobFuncs[] =
 
 // ============================================================================
 
-static const asglobproperties_t asGlobProps[] =
+static const gs_asglobproperties_t asGlobProps[] =
 {
 	{ "const int64 levelTime", &level.time },
 	{ "const uint frameTime", &game.frametime },
@@ -3145,54 +2377,6 @@ static const asglobproperties_t asGlobProps[] =
 
 	{ NULL }
 };
-
-static void G_asRegisterGlobalFunctions( asIScriptEngine *asEngine, const asglobfuncs_t *funcs, const char *nameSpace ) {
-	int error;
-	int count = 0, failedcount = 0;
-	const asglobfuncs_t *func;
-
-	asEngine->SetDefaultNamespace( nameSpace );
-
-	for( func = funcs; func->declaration; func++ ) {
-		error = asEngine->RegisterGlobalFunction( func->declaration, func->pointer, asCALL_CDECL );
-
-		if( error < 0 ) {
-			failedcount++;
-			continue;
-		}
-
-		count++;
-	}
-
-	// get AS function pointers
-	for( func = funcs; func->declaration; func++ ) {
-		if( func->asFuncPtr ) {
-			*func->asFuncPtr = asEngine->GetGlobalFunctionByDecl( func->declaration );
-		}
-	}
-
-	asEngine->SetDefaultNamespace( "" );
-}
-
-static void G_asRegisterGlobalProperties( asIScriptEngine *asEngine, const asglobproperties_t *props, const char *nameSpace ) {
-	int error;
-	int count = 0, failedcount = 0;
-	const asglobproperties_t *prop;
-
-	asEngine->SetDefaultNamespace( nameSpace );
-
-	for( prop = props; prop->declaration; prop++ ) {
-		error = asEngine->RegisterGlobalProperty( prop->declaration, prop->pointer );
-		if( error < 0 ) {
-			failedcount++;
-			continue;
-		}
-
-		count++;
-	}
-
-	asEngine->SetDefaultNamespace( "" );
-}
 
 // ==========================================================================================
 
@@ -3231,7 +2415,7 @@ bool G_asCallMapEntitySpawnScript( const char *classname, edict_t *ent ) {
 	G_asClearEntityBehaviors( ent );
 
 	// call the spawn function
-	asContext = angelExport->asAcquireContext( asEngine );
+	asContext = game.asExport->asAcquireContext( asEngine );
 	error = asContext->Prepare( asSpawnFunc );
 	if( error < 0 ) {
 		return false;
@@ -3284,7 +2468,7 @@ void G_asClearEntityBehaviors( edict_t *ent ) {
 * Release callback function references held by the engine for script spawned entities
 */
 void G_asReleaseEntityBehaviors( edict_t *ent ) {
-	if( ent->scriptSpawned && angelExport ) {
+	if( ent->scriptSpawned && game.asExport ) {
 		if( ent->asThinkFunc ) {
 			static_cast<asIScriptFunction*>( ent->asThinkFunc )->Release();
 		}
@@ -3317,7 +2501,7 @@ void G_asCallMapEntityThink( edict_t *ent ) {
 		return;
 	}
 
-	ctx = angelExport->asAcquireContext( GAME_AS_ENGINE() );
+	ctx = game.asExport->asAcquireContext( GAME_AS_ENGINE() );
 
 	error = ctx->Prepare( static_cast<asIScriptFunction *>( ent->asThinkFunc ) );
 	if( error < 0 ) {
@@ -3343,7 +2527,7 @@ void G_asCallMapEntityTouch( edict_t *ent, edict_t *other, cplane_t *plane, int 
 		return;
 	}
 
-	ctx = angelExport->asAcquireContext( GAME_AS_ENGINE() );
+	ctx = game.asExport->asAcquireContext( GAME_AS_ENGINE() );
 
 	error = ctx->Prepare( static_cast<asIScriptFunction *>( ent->asTouchFunc ) );
 	if( error < 0 ) {
@@ -3377,7 +2561,7 @@ void G_asCallMapEntityUse( edict_t *ent, edict_t *other, edict_t *activator ) {
 		return;
 	}
 
-	ctx = angelExport->asAcquireContext( GAME_AS_ENGINE() );
+	ctx = game.asExport->asAcquireContext( GAME_AS_ENGINE() );
 
 	error = ctx->Prepare( static_cast<asIScriptFunction *>( ent->asUseFunc ) );
 	if( error < 0 ) {
@@ -3404,7 +2588,7 @@ void G_asCallMapEntityPain( edict_t *ent, edict_t *other, float kick, float dama
 		return;
 	}
 
-	ctx = angelExport->asAcquireContext( GAME_AS_ENGINE() );
+	ctx = game.asExport->asAcquireContext( GAME_AS_ENGINE() );
 
 	error = ctx->Prepare( static_cast<asIScriptFunction *>( ent->asPainFunc ) );
 	if( error < 0 ) {
@@ -3432,7 +2616,7 @@ void G_asCallMapEntityDie( edict_t *ent, edict_t *inflicter, edict_t *attacker, 
 		return;
 	}
 
-	ctx = angelExport->asAcquireContext( GAME_AS_ENGINE() );
+	ctx = game.asExport->asAcquireContext( GAME_AS_ENGINE() );
 
 	error = ctx->Prepare( static_cast<asIScriptFunction *>( ent->asDieFunc ) );
 	if( error < 0 ) {
@@ -3459,7 +2643,7 @@ void G_asCallMapEntityStop( edict_t *ent ) {
 		return;
 	}
 
-	ctx = angelExport->asAcquireContext( GAME_AS_ENGINE() );
+	ctx = game.asExport->asAcquireContext( GAME_AS_ENGINE() );
 
 	error = ctx->Prepare( static_cast<asIScriptFunction *>( ent->asStopFunc ) );
 	if( error < 0 ) {
@@ -3491,7 +2675,7 @@ bool G_ExecutionErrorReport( int error ) {
 * G_LoadGameScript
 */
 asIScriptModule *G_LoadGameScript( const char *moduleName, const char *dir, const char *filename, const char *ext ) {
-	return angelExport->asLoadScriptProject( GAME_AS_ENGINE(), moduleName, GAME_SCRIPTS_DIRECTORY, dir, filename, ext );
+	return game.asExport->asLoadScriptProject( GAME_AS_ENGINE(), moduleName, GAME_SCRIPTS_DIRECTORY, dir, filename, ext );
 }
 
 /*
@@ -3521,24 +2705,27 @@ static void G_InitializeGameModuleSyntax( asIScriptEngine *asEngine ) {
 
 	G_Printf( "* Initializing Game module syntax\n" );
 
+	// register shared stuff
+	GS_asInitializeEngine( asEngine );
+
 	// register global variables
-	G_asRegisterEnums( asEngine, asGameEnums );
-	G_asRegisterEnums( asEngine, asAIEnums );
+	GS_asRegisterEnums( asEngine, asGameEnums );
+	GS_asRegisterEnums( asEngine, asAIEnums );
 
 	// first register all class names so methods using custom classes work
-	G_asRegisterObjectClassNames( asEngine, asGameClassesDescriptors );
-	G_asRegisterObjectClassNames( asEngine, asAIClassesDescriptors );
+	GS_asRegisterObjectClassNames( asEngine, asGameClassesDescriptors );
+	GS_asRegisterObjectClassNames( asEngine, asAIClassesDescriptors );
 
 	// register classes
-	G_asRegisterObjectClasses( asEngine, asGameClassesDescriptors );
-	G_asRegisterObjectClasses( asEngine, asAIClassesDescriptors );
+	GS_asRegisterObjectClasses( asEngine, asGameClassesDescriptors );
+	GS_asRegisterObjectClasses( asEngine, asAIClassesDescriptors );
 
 	// register global functions
-	G_asRegisterGlobalFunctions( asEngine, asGameGlobFuncs, "" );
-	G_asRegisterGlobalFunctions( asEngine, asAIGlobFuncs, "AI" );
+	GS_asRegisterGlobalFunctions( asEngine, asGameGlobFuncs, "" );
+	GS_asRegisterGlobalFunctions( asEngine, asAIGlobFuncs, "AI" );
 
 	// register global properties
-	G_asRegisterGlobalProperties( asEngine, asGlobProps, "" );
+	GS_asRegisterGlobalProperties( asEngine, asGlobProps, "" );
 }
 
 /*
@@ -3551,13 +2738,13 @@ void G_asInitGameModuleEngine( void ) {
 	G_ResetGameModuleScriptData();
 
 	// initialize the engine
-	angelExport = trap_asGetAngelExport();
-	if( !angelExport ) {
+	game.asExport = trap_asGetAngelExport();
+	if( !game.asExport ) {
 		G_Printf( "* Couldn't initialize angelscript, missing symbol.\n" );
 		return;
 	}
 
-	asEngine = angelExport->asCreateEngine( &asGeneric );
+	asEngine = game.asExport->asCreateEngine( &asGeneric );
 	if( !asEngine ) {
 		G_Printf( "* Couldn't initialize angelscript.\n" );
 		return;
@@ -3579,8 +2766,8 @@ void G_asInitGameModuleEngine( void ) {
 */
 void G_asShutdownGameModuleEngine( void ) {
 	if( game.asEngine != NULL ) {
-		if( angelExport ) {
-			angelExport->asReleaseEngine( static_cast<asIScriptEngine *>( game.asEngine ) );
+		if( game.asExport ) {
+			game.asExport->asReleaseEngine( static_cast<asIScriptEngine *>( game.asEngine ) );
 		}
 		G_ResetGameModuleScriptData();
 	}
@@ -3596,7 +2783,7 @@ void G_asGarbageCollect( bool force ) {
 	unsigned int currentSize, totalDestroyed, totalDetected;
 	asIScriptEngine *asEngine;
 
-	if( !angelExport ) {
+	if( !game.asExport ) {
 		return;
 	}
 
@@ -3630,15 +2817,15 @@ void G_asGarbageCollect( bool force ) {
 static void G_asDumpAPIToFile( const char *path ) {
 	int i, j;
 	int file;
-	const asClassDescriptor_t *cDescr;
+	const gs_asClassDescriptor_t *cDescr;
 	const char *name;
 	char *filename = NULL;
 	size_t filename_size = 0;
 	char string[1024];
 
 	// dump class definitions, containing methods, behaviors and properties
-	const asClassDescriptor_t *const *allDescriptors[] = { asGameClassesDescriptors, asAIClassesDescriptors };
-	for( const asClassDescriptor_t *const *descriptors: allDescriptors ) {
+	const gs_asClassDescriptor_t *const *allDescriptors[] = { asGameClassesDescriptors, asAIClassesDescriptors };
+	for( const gs_asClassDescriptor_t *const *descriptors: allDescriptors ) {
 		for( i = 0;; i++ ) {
 			if( !( cDescr = descriptors[i] ) ) {
 				break;
@@ -3665,7 +2852,7 @@ static void G_asDumpAPIToFile( const char *path ) {
 				trap_FS_Write( string, strlen( string ), file );
 
 				for( j = 0;; j++ ) {
-					const asFuncdef_t *funcdef = &cDescr->funcdefs[j];
+					const gs_asFuncdef_t *funcdef = &cDescr->funcdefs[j];
 					if( !funcdef->declaration ) {
 						break;
 					}
@@ -3690,7 +2877,7 @@ static void G_asDumpAPIToFile( const char *path ) {
 				trap_FS_Write( string, strlen( string ), file );
 
 				for( j = 0;; j++ ) {
-					const asProperty_t *objProperty = &cDescr->objProperties[j];
+					const gs_asProperty_t *objProperty = &cDescr->objProperties[j];
 					if( !objProperty->declaration ) {
 						break;
 					}
@@ -3706,7 +2893,7 @@ static void G_asDumpAPIToFile( const char *path ) {
 				trap_FS_Write( string, strlen( string ), file );
 
 				for( j = 0;; j++ ) {
-					const asBehavior_t *objBehavior = &cDescr->objBehaviors[j];
+					const gs_asBehavior_t *objBehavior = &cDescr->objBehaviors[j];
 					if( !objBehavior->declaration ) {
 						break;
 					}
@@ -3729,7 +2916,7 @@ static void G_asDumpAPIToFile( const char *path ) {
 				trap_FS_Write( string, strlen( string ), file );
 
 				for( j = 0;; j++ ) {
-					const asMethod_t *objMethod = &cDescr->objMethods[j];
+					const gs_asMethod_t *objMethod = &cDescr->objMethods[j];
 					if( !objMethod->declaration ) {
 						break;
 					}
@@ -3766,14 +2953,14 @@ static void G_asDumpAPIToFile( const char *path ) {
 
 	// enums
 	{
-		const asEnum_t *asEnum;
-		const asEnumVal_t *asEnumVal;
+		const gs_asEnum_t *asEnum;
+		const gs_asEnumVal_t *asEnumVal;
 
 		Q_snprintfz( string, sizeof( string ), "/**\r\n * %s\r\n */\r\n", "Enums" );
 		trap_FS_Write( string, strlen( string ), file );
 
-		const asEnum_t *const allEnumsLists[] = { asGameEnums, asAIEnums };
-		for( const asEnum_t *const enumsList: allEnumsLists ) {
+		const gs_asEnum_t *const allEnumsLists[] = { asGameEnums, asAIEnums };
+		for( const gs_asEnum_t *const enumsList: allEnumsLists ) {
 			for( i = 0, asEnum = enumsList; asEnum->name != NULL; i++, asEnum++ ) {
 				Q_snprintfz( string, sizeof( string ), "typedef enum\r\n{\r\n" );
 				trap_FS_Write( string, strlen( string ), file );
@@ -3791,7 +2978,7 @@ static void G_asDumpAPIToFile( const char *path ) {
 
 	// global properties
 	{
-		const asglobproperties_t *prop;
+		const gs_asglobproperties_t *prop;
 
 		Q_snprintfz( string, sizeof( string ), "/**\r\n * %s\r\n */\r\n", "Global properties" );
 		trap_FS_Write( string, strlen( string ), file );
@@ -3807,13 +2994,13 @@ static void G_asDumpAPIToFile( const char *path ) {
 
 	// global functions
 	{
-		const asglobfuncs_t *func;
+		const gs_asglobfuncs_t *func;
 
 		Q_snprintfz( string, sizeof( string ), "/**\r\n * %s\r\n */\r\n", "Global functions" );
 		trap_FS_Write( string, strlen( string ), file );
 
-		const asglobfuncs_t *const allFuncsList[] = { asGameGlobFuncs, asAIGlobFuncs };
-		for( const asglobfuncs_t *funcsList: allFuncsList ) {
+		const gs_asglobfuncs_t *const allFuncsList[] = { asGameGlobFuncs, asAIGlobFuncs };
+		for( const gs_asglobfuncs_t *funcsList: allFuncsList ) {
 			for( func = funcsList; func->declaration; func++ ) {
 				Q_snprintfz( string, sizeof( string ), "%s;\r\n", func->declaration );
 				trap_FS_Write( string, strlen( string ), file );
