+++ source/gameshared/gs_misc.c
@@ -25,25 +25,10 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "q_collision.h"
 #include "gs_public.h"
 
-#ifndef _MSC_VER
-void ( *module_Printf )( const char *format, ... ) __attribute__( ( format( printf, 1, 2 ) ) );
-void ( *module_Error)( const char *format, ... ) __attribute__( ( format( printf, 1, 2 ) ) ) __attribute__( ( noreturn ) );
-#else
-void ( *module_Printf )( const char *format, ... );
-void ( *module_Error )( const char *format, ... );
-#endif
-
-void *( *module_Malloc )( size_t size );
-void ( *module_Free )( void *data );
-void ( *module_Trace )( trace_t *t, vec3_t start, vec3_t mins, vec3_t maxs, vec3_t end, int ignore, int contentmask, int timeDelta );
-entity_state_t *( *module_GetEntityState )( int entNum, int deltaTime );
-int ( *module_PointContents )( vec3_t point, int timeDelta );
-void ( *module_PredictedEvent )( int entNum, int ev, int parm );
-void ( *module_PMoveTouchTriggers )( pmove_t *pm, vec3_t previous_origin );
-const char *( *module_GetConfigString )( int index );
-
 // TEMP MOVE ME
 gs_state_t gs;
+
+void GS_asInitializeExport( void );
 
 /*
 * GS_TouchPushTrigger
@@ -60,7 +45,7 @@ void GS_TouchPushTrigger( player_state_t *playerState, entity_state_t *pusher ) 
 	playerState->pmove.pm_flags &= ~PMF_WALLJUMPCOUNT;
 	playerState->pmove.pm_flags |= PMF_JUMPPAD_TIME;
 	playerState->pmove.pm_flags &= ~PMF_ON_GROUND;
-	module_PredictedEvent( playerState->POVnum, EV_JUMP_PAD, 0 );
+	gs.api.PredictedEvent( playerState->POVnum, EV_JUMP_PAD, 0 );
 }
 
 /*
@@ -76,15 +61,15 @@ int GS_WaterLevel( entity_state_t *state, vec3_t mins, vec3_t maxs ) {
 	point[0] = state->origin[0];
 	point[1] = state->origin[1];
 	point[2] = state->origin[2] + mins[2] + 1;
-	cont = module_PointContents( point, 0 );
+	cont = gs.api.PointContents( point, 0 );
 	if( cont & MASK_WATER ) {
 		waterlevel = 1;
 		point[2] += 26;
-		cont = module_PointContents( point, 0 );
+		cont = gs.api.PointContents( point, 0 );
 		if( cont & MASK_WATER ) {
 			waterlevel = 2;
 			point[2] += 22;
-			cont = module_PointContents( point, 0 );
+			cont = gs.api.PointContents( point, 0 );
 			if( cont & MASK_WATER ) {
 				waterlevel = 3;
 			}
@@ -102,7 +87,7 @@ void GS_BBoxForEntityState( entity_state_t *state, vec3_t mins, vec3_t maxs ) {
 
 	if( state->solid == SOLID_BMODEL ) {
 		// FIXME: This is wrong, we don't have access to bmodels at gameshared (simply didn't add it)
-		module_Error( "GS_BBoxForEntityState: called for a brush model\n" );
+		gs.api.Error( "GS_BBoxForEntityState: called for a brush model\n" );
 		//cmodel = trap_CM_InlineModel( state->modelindex );
 	} else {                          // encoded bbox
 		x = 8 * ( state->solid & 31 );
@@ -300,4 +285,17 @@ void GS_Obituary( void *victim, int gender, void *attacker, int mod, char *messa
 			strcpy( message, "was fragged by" );
 			break;
 	}
+}
+
+/*
+* GS_InitModule
+*/
+void GS_InitModule( int module, int maxClients, gs_module_api_t *api ) {
+	memset( &gs, 0, sizeof( gs_state_t ) );
+
+	gs.module = module;
+	gs.maxclients = maxClients;
+	gs.api = *api;
+
+	GS_asInitializeExport();
 }
