+++ source/gameshared/gs_pmove.c
@@ -171,7 +171,7 @@ static void PlayerTouchWall( int nbTestDir, float maxZnormal, vec3_t *normal ) {
 		mins[1] += pml.velocity[1] * 0.015f;
 	}
 	mins[2] = maxs[2] = 0;
-	module_Trace( &trace, pml.origin, mins, maxs, pml.origin, pm->playerState->POVnum, pm->contentmask, 0 );
+	gs.api.Trace( &trace, pml.origin, mins, maxs, pml.origin, pm->playerState->POVnum, pm->contentmask, 0 );
 	if( !trace.allsolid && trace.fraction == 1 ) {
 		return;
 	}
@@ -226,7 +226,7 @@ static void PlayerTouchWall( int nbTestDir, float maxZnormal, vec3_t *normal ) {
 		dir[1] = pml.origin[1] + dy * m + pml.velocity[1] * 0.015f;
 		dir[2] = pml.origin[2];
 
-		module_Trace( &trace, pml.origin, zero, zero, dir, pm->playerState->POVnum, pm->contentmask, 0 );
+		gs.api.Trace( &trace, pml.origin, zero, zero, dir, pm->playerState->POVnum, pm->contentmask, 0 );
 
 		if( trace.allsolid ) {
 			return;
@@ -240,7 +240,7 @@ static void PlayerTouchWall( int nbTestDir, float maxZnormal, vec3_t *normal ) {
 			continue;
 		}
 
-		if( trace.ent > 0 && module_GetEntityState( trace.ent, 0 )->type == ET_PLAYER ) {
+		if( trace.ent > 0 && gs.api.GetEntityState( trace.ent, 0 )->type == ET_PLAYER ) {
 			continue;
 		}
 
@@ -316,7 +316,7 @@ static int PM_SlideMove( void ) {
 
 	for( moves = 0; moves < maxmoves; moves++ ) {
 		VectorMA( pml.origin, remainingTime, pml.velocity, end );
-		module_Trace( &trace, pml.origin, pm->mins, pm->maxs, end, pm->playerState->POVnum, pm->contentmask, 0 );
+		gs.api.Trace( &trace, pml.origin, pm->mins, pm->maxs, end, pm->playerState->POVnum, pm->contentmask, 0 );
 		if( trace.allsolid ) { // trapped into a solid
 			VectorCopy( last_valid_origin, pml.origin );
 			return SLIDEMOVEFLAG_TRAPPED;
@@ -462,7 +462,7 @@ static void PM_StepSlideMove( void ) {
 	VectorCopy( start_o, up );
 	up[2] += STEPSIZE;
 
-	module_Trace( &trace, up, pm->mins, pm->maxs, up, pm->playerState->POVnum, pm->contentmask, 0 );
+	gs.api.Trace( &trace, up, pm->mins, pm->maxs, up, pm->playerState->POVnum, pm->contentmask, 0 );
 	if( trace.allsolid ) {
 		return; // can't step up
 
@@ -476,7 +476,7 @@ static void PM_StepSlideMove( void ) {
 	// push down the final amount
 	VectorCopy( pml.origin, down );
 	down[2] -= STEPSIZE;
-	module_Trace( &trace, pml.origin, pm->mins, pm->maxs, down, pm->playerState->POVnum, pm->contentmask, 0 );
+	gs.api.Trace( &trace, pml.origin, pm->mins, pm->maxs, down, pm->playerState->POVnum, pm->contentmask, 0 );
 	if( !trace.allsolid ) {
 		VectorCopy( trace.endpos, pml.origin );
 	}
@@ -974,7 +974,7 @@ static void PM_GroundTrace( trace_t *trace ) {
 	point[1] = pml.origin[1];
 	point[2] = pml.origin[2] - 0.25;
 
-	module_Trace( trace, pml.origin, pm->mins, pm->maxs, point, pm->playerState->POVnum, pm->contentmask, 0 );
+	gs.api.Trace( trace, pml.origin, pm->mins, pm->maxs, point, pm->playerState->POVnum, pm->contentmask, 0 );
 }
 
 /*
@@ -985,7 +985,7 @@ static bool PM_GoodPosition( vec3_t origin, trace_t *trace ) {
 		return true;
 	}
 
-	module_Trace( trace, origin, pm->mins, pm->maxs, origin, pm->playerState->POVnum, pm->contentmask, 0 );
+	gs.api.Trace( trace, origin, pm->mins, pm->maxs, origin, pm->playerState->POVnum, pm->contentmask, 0 );
 
 	return !trace->allsolid;
 }
@@ -1083,17 +1083,17 @@ static void PM_CategorizePosition( void ) {
 	point[0] = pml.origin[0];
 	point[1] = pml.origin[1];
 	point[2] = pml.origin[2] + pm->mins[2] + 1;
-	cont = module_PointContents( point, 0 );
+	cont = gs.api.PointContents( point, 0 );
 
 	if( cont & MASK_WATER ) {
 		pm->watertype = cont;
 		pm->waterlevel = 1;
 		point[2] = pml.origin[2] + pm->mins[2] + sample1;
-		cont = module_PointContents( point, 0 );
+		cont = gs.api.PointContents( point, 0 );
 		if( cont & MASK_WATER ) {
 			pm->waterlevel = 2;
 			point[2] = pml.origin[2] + pm->mins[2] + sample2;
-			cont = module_PointContents( point, 0 );
+			cont = gs.api.PointContents( point, 0 );
 			if( cont & MASK_WATER ) {
 				pm->waterlevel = 3;
 			}
@@ -1167,13 +1167,13 @@ static void PM_CheckJump( void ) {
 
 	//if( gs.module == GS_MODULE_GAME ) GS_Printf( "upvel %f\n", pml.velocity[2] );
 	if( pml.velocity[2] > 100 ) {
-		module_PredictedEvent( pm->playerState->POVnum, EV_DOUBLEJUMP, 0 );
+		gs.api.PredictedEvent( pm->playerState->POVnum, EV_DOUBLEJUMP, 0 );
 		pml.velocity[2] += pml.jumpPlayerSpeed;
 	} else if( pml.velocity[2] > 0 ) {
-		module_PredictedEvent( pm->playerState->POVnum, EV_JUMP, 0 );
+		gs.api.PredictedEvent( pm->playerState->POVnum, EV_JUMP, 0 );
 		pml.velocity[2] += pml.jumpPlayerSpeed;
 	} else {
-		module_PredictedEvent( pm->playerState->POVnum, EV_JUMP, 0 );
+		gs.api.PredictedEvent( pm->playerState->POVnum, EV_JUMP, 0 );
 		pml.velocity[2] = pml.jumpPlayerSpeed;
 	}
 
@@ -1259,14 +1259,14 @@ static void PM_CheckDash( void ) {
 		// return sound events
 		if( fabs( pml.sidePush ) > 10 && fabs( pml.sidePush ) >= fabs( pml.forwardPush ) ) {
 			if( pml.sidePush > 0 ) {
-				module_PredictedEvent( pm->playerState->POVnum, EV_DASH, 2 );
+				gs.api.PredictedEvent( pm->playerState->POVnum, EV_DASH, 2 );
 			} else {
-				module_PredictedEvent( pm->playerState->POVnum, EV_DASH, 1 );
+				gs.api.PredictedEvent( pm->playerState->POVnum, EV_DASH, 1 );
 			}
 		} else if( pml.forwardPush < -10 ) {
-			module_PredictedEvent( pm->playerState->POVnum, EV_DASH, 3 );
+			gs.api.PredictedEvent( pm->playerState->POVnum, EV_DASH, 3 );
 		} else {
-			module_PredictedEvent( pm->playerState->POVnum, EV_DASH, 0 );
+			gs.api.PredictedEvent( pm->playerState->POVnum, EV_DASH, 0 );
 		}
 	} else if( pm->groundentity == -1 ) {
 		pm->playerState->pmove.pm_flags &= ~PMF_DASHING;
@@ -1326,7 +1326,7 @@ static void PM_CheckWallJump( void ) {
 		// don't walljump if our height is smaller than a step
 		// unless jump is pressed or the player is moving faster than dash speed and upwards
 		hspeed = VectorLengthFast( tv( pml.velocity[0], pml.velocity[1], 0 ) );
-		module_Trace( &trace, pml.origin, pm->mins, pm->maxs, point, pm->playerState->POVnum, pm->contentmask, 0 );
+		gs.api.Trace( &trace, pml.origin, pm->mins, pm->maxs, point, pm->playerState->POVnum, pm->contentmask, 0 );
 
 		if( pml.upPush >= 10
 			|| ( hspeed > pm->playerState->pmove.stats[PM_STAT_DASHSPEED] && pml.velocity[2] > 8 )
@@ -1380,13 +1380,13 @@ static void PM_CheckWallJump( void ) {
 					pm->playerState->pmove.stats[PM_STAT_WJTIME] = PM_WALLJUMP_FAILED_TIMEDELAY;
 
 					// Create the event
-					module_PredictedEvent( pm->playerState->POVnum, EV_WALLJUMP_FAILED, DirToByte( normal ) );
+					gs.api.PredictedEvent( pm->playerState->POVnum, EV_WALLJUMP_FAILED, DirToByte( normal ) );
 				} else {
 					pm->playerState->pmove.stats[PM_STAT_WJTIME] = PM_WALLJUMP_TIMEDELAY;
 					pm->playerState->pmove.skim_time = PM_SKIM_TIME;
 
 					// Create the event
-					module_PredictedEvent( pm->playerState->POVnum, EV_WALLJUMP, DirToByte( normal ) );
+					gs.api.PredictedEvent( pm->playerState->POVnum, EV_WALLJUMP, DirToByte( normal ) );
 				}
 			}
 		}
@@ -1439,7 +1439,7 @@ static void PM_CheckSpecialMovement( void ) {
 	// check for ladder
 	if( !pm->skipCollision ) {
 		VectorMA( pml.origin, 1, pml.flatforward, spot );
-		module_Trace( &trace, pml.origin, pm->mins, pm->maxs, spot, pm->playerState->POVnum, pm->contentmask, 0 );
+		gs.api.Trace( &trace, pml.origin, pm->mins, pm->maxs, spot, pm->playerState->POVnum, pm->contentmask, 0 );
 		if( ( trace.fraction < 1 ) && ( trace.surfFlags & SURF_LADDER ) ) {
 			pml.ladder = true;
 			pm->ladder = true;
@@ -1453,13 +1453,13 @@ static void PM_CheckSpecialMovement( void ) {
 
 	VectorMA( pml.origin, 30, pml.flatforward, spot );
 	spot[2] += 4;
-	cont = module_PointContents( spot, 0 );
+	cont = gs.api.PointContents( spot, 0 );
 	if( !( cont & CONTENTS_SOLID ) ) {
 		return;
 	}
 
 	spot[2] += 16;
-	cont = module_PointContents( spot, 0 );
+	cont = gs.api.PointContents( spot, 0 );
 	if( cont ) {
 		return;
 	}
@@ -1556,7 +1556,7 @@ static void PM_FlyMove( bool doclip ) {
 		for( i = 0; i < 3; i++ )
 			end[i] = pml.origin[i] + pml.frametime * pml.velocity[i];
 
-		module_Trace( &trace, pml.origin, pm->mins, pm->maxs, end, pm->playerState->POVnum, pm->contentmask, 0 );
+		gs.api.Trace( &trace, pml.origin, pm->mins, pm->maxs, end, pm->playerState->POVnum, pm->contentmask, 0 );
 
 		VectorCopy( trace.endpos, pml.origin );
 	} else {
@@ -1651,7 +1651,7 @@ static void PM_AdjustBBox( void ) {
 		wishviewheight = playerbox_stand_viewheight - ( crouchFrac * ( playerbox_stand_viewheight - playerbox_crouch_viewheight ) );
 
 		// check that the head is not blocked
-		module_Trace( &trace, pml.origin, wishmins, wishmaxs, pml.origin, pm->playerState->POVnum, pm->contentmask, 0 );
+		gs.api.Trace( &trace, pml.origin, wishmins, wishmaxs, pml.origin, pm->playerState->POVnum, pm->contentmask, 0 );
 		if( trace.allsolid || trace.startsolid ) {
 			// can't do the uncrouching, let the time alone and use old position
 			VectorCopy( curmins, pm->mins );
@@ -2054,7 +2054,7 @@ void Pmove( pmove_t *pmove ) {
 	// We check the entire path between the origin before the pmove and the
 	// current origin to ensure no triggers are missed at high velocity.
 	// Note that this method assumes the movement has been linear.
-	module_PMoveTouchTriggers( pm, pml.previous_origin );
+	gs.api.PMoveTouchTriggers( pm, pml.previous_origin );
 
 	PM_UpdateDeltaAngles(); // in case some trigger action has moved the view angles (like teleported).
 
@@ -2097,7 +2097,7 @@ void Pmove( pmove_t *pmove ) {
 				clamp( damage, 0.0f, MAX_FALLING_DAMAGE );
 			}
 
-			module_PredictedEvent( pm->playerState->POVnum, EV_FALL, damage );
+			gs.api.PredictedEvent( pm->playerState->POVnum, EV_FALL, damage );
 		}
 
 		pm->playerState->pmove.pm_flags &= ~PMF_JUMPPAD_TIME;
