+++ source/gameshared/gs_slidebox.c
@@ -184,7 +184,7 @@ static void GS_AddClippingPlane( move_t *move, const vec3_t planeNormal ) {
 	}
 
 	if( move->numClipPlanes + 1 == MAX_SLIDEMOVE_CLIP_PLANES ) {
-		module_Error( "GS_AddTouchPlane: MAX_SLIDEMOVE_CLIP_PLANES reached\n" );
+		gs.api.Error( "GS_AddTouchPlane: MAX_SLIDEMOVE_CLIP_PLANES reached\n" );
 	}
 
 	// add the plane
@@ -201,7 +201,7 @@ static int GS_SlideMoveClipMove( move_t *move /*, const bool stepping*/ ) {
 	int blockedmask = 0;
 
 	VectorMA( move->origin, move->remainingTime, move->velocity, endpos );
-	module_Trace( &trace, move->origin, move->mins, move->maxs, endpos, move->passent, move->contentmask, 0 );
+	gs.api.Trace( &trace, move->origin, move->mins, move->maxs, endpos, move->passent, move->contentmask, 0 );
 	if( trace.allsolid ) {
 		if( trace.ent > 0 ) {
 			GS_AddTouchEnt( move, trace.ent );
@@ -273,7 +273,7 @@ int GS_SlideMove( move_t *move ) {
 #ifdef CHECK_TRAPPED
 		{
 			trace_t trace;
-			module_Trace( &trace, move->origin, move->mins, move->maxs, move->origin, move->passent, move->contentmask, 0 );
+			gs.api.Trace( &trace, move->origin, move->mins, move->maxs, move->origin, move->passent, move->contentmask, 0 );
 			if( trace.startsolid ) {
 				blockedmask |= SLIDEMOVEFLAG_TRAPPED;
 			}
@@ -283,7 +283,7 @@ int GS_SlideMove( move_t *move ) {
 		// can't continue
 		if( blockedmask & SLIDEMOVEFLAG_TRAPPED ) {
 #ifdef CHECK_TRAPPED
-			module_Printf( "GS_SlideMove SLIDEMOVEFLAG_TRAPPED\n" );
+			gs.api.Printf( "GS_SlideMove SLIDEMOVEFLAG_TRAPPED\n" );
 #endif
 			move->remainingTime = 0.0f;
 			VectorCopy( lastValidOrigin, move->origin );
@@ -299,7 +299,7 @@ int GS_SlideMove( move_t *move ) {
 
 		// if it didn't touch anything the move should be completed
 		if( move->remainingTime > 0.0f ) {
-			module_Printf( "slidemove finished with remaining time\n" );
+			gs.api.Printf( "slidemove finished with remaining time\n" );
 			move->remainingTime = 0.0f;
 		}
